% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{abs.ComplexLazyTensor}
\alias{abs.ComplexLazyTensor}
\title{Element-wise absolute value.}
\usage{
\method{abs}{ComplexLazyTensor}(x)
}
\description{
Element-wise absolute value.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
