% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{abs.LazyTensor}
\alias{abs.LazyTensor}
\alias{abs.ComplexLazyTensor}
\title{Element-wise absolute value operation}
\usage{
\method{abs}{LazyTensor}(x)

\method{abs}{ComplexLazyTensor}(x)
}
\arguments{
\item{x}{a \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values,
or a scalar value.}
}
\value{
An object of class \code{LazyTensor}.
}
\description{
Symbolic element-wise absolute value (or modulus) operation for
\code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{abs(x)} returns a \code{LazyTensor} that
encodes, symbolically, the element-wise absolute value of \code{x}.

If \code{x} is a \code{ComplexLazyTensor}, \code{abs(x)} returns a \code{LazyTensor} that
encodes, symbolically, the complex absolute value or modulus of \code{x}, i.e.
\code{abs(x) = Mod(x)}.
}
\examples{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
                                    
Abs_x <- abs(x_i)                   # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=abs]{abs()}}, \code{\link[=Mod]{Mod()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
