% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Mod.ComplexLazyTensor}
\alias{Mod.ComplexLazyTensor}
\title{Element-wise absolute value (or modulus).}
\usage{
\method{Mod}{ComplexLazyTensor}(z)
}
\description{
Element-wise absolute value (or modulus).
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
