% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Mod.default}
\alias{Mod.default}
\title{Absolute value (or modulus).}
\usage{
\method{Mod}{default}()
}
\description{
Absolute value (or modulus).
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
