% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Arg.ComplexLazyTensor}
\alias{Arg.ComplexLazyTensor}
\title{Element-wise angle (or argument) of complex.}
\usage{
\method{Arg}{ComplexLazyTensor}(z)
}
\description{
Element-wise angle (or argument) of complex.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
