% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Arg.default}
\alias{Arg.default}
\title{Angle (or argument) of complex.}
\usage{
\method{Arg}{default}()
}
\description{
Angle (or argument) of complex.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
