% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{argKmin}
\alias{argKmin}
\alias{argKmin_reduction}
\title{Arg-K-min reduction}
\usage{
argKmin(x, K, index)

argKmin_reduction(x, K, index)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{K}{An \code{integer} corresponding to the  number of minimal values
required.}

\item{index}{a \code{character} corresponding to the reduction dimension that
should be either \code{"i"} or \code{"j"} to specify whether if the reduction is
indexed by \code{"i"} or \code{"j"}.}
}
\value{
A matrix corresponding to the argKmin reduction.
}
\description{
Arg-K-min reduction for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor} or a \code{ComplexLazyTensor},
\code{argKmin(x, K, index)} will:
\itemize{
\item if \code{index = "i"}, return the indices of the \code{K} minimal values of \code{x}
over the \code{i} indexes.
\item if \code{index = "j"}, return the indices of the \code{K} minimal values of \code{x}
over the \code{j} indexes.
}

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Reductions"}.
}
\examples{
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
K <- 2
argkmin_x <- argKmin(x_i, K, "i")   # argKmin reduction 
                                    # indexed by 'i'

\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
K <- 2
argkmin_red_x <- argKmin_reduction(x_i, K, "i")  # argKmin reduction 
                                                 # indexed by 'i'
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
