% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{*.ComplexLazyTensor}
\alias{*.ComplexLazyTensor}
\title{Multiplication.}
\usage{
\method{*}{ComplexLazyTensor}(x, y)
}
\description{
Multiplication.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
