% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{asin}
\alias{asin}
\title{Element-wise arc-sine operation.}
\usage{
asin(x)
}
\arguments{
\item{x}{input for \code{\link[=asin.default]{asin.default()}} or
\code{\link[=asin.LazyTensor]{asin.LazyTensor()}}.}
}
\value{
See value of \code{\link[=asin.default]{asin.default()}} or
\code{\link[=asin.LazyTensor]{asin.LazyTensor()}}.
}
\description{
Symbolic element-wise arc-sine operation for \code{LazyTensor}
objects or standard element-wise arc-sine operation otherwise.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=asin.LazyTensor]{asin.LazyTensor()}}, else
see \code{\link[=asin.default]{asin.default()}}.
}
\examples{
# R base operation
asin(0)
asin(-1)
# LazyTensor symbolic element-wise arc-sine
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Asin_x <- asin(x_i)                 # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=asin.default]{asin.default()}}, \code{\link[=asin.LazyTensor]{asin.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
