% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{atan2.default}
\alias{atan2.default}
\title{2-argument arc-tangent.}
\usage{
\method{atan2}{default}(x, y)
}
\description{
2-argument arc-tangent.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
