% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-check.R
\name{check_cmake}
\alias{check_cmake}
\title{Check cmake version}
\usage{
check_cmake(cmake_executable, onLoad = FALSE)
}
\arguments{
\item{cmake_executable}{text string, path to cmake
(e.g. \verb{/path/to/cmake} on Unix system).}
}
\value{
a value 1 or 0 depending if checks are ok or not.
}
\description{
This function calls the cmake executable whose path is
given in input and check if its version is higher than
3.10 which is the minimum required for \code{rkeops} to work.
}
\details{
The function check for the existence of the input file
\code{cmake_executable}, if it is actually a cmake executable
and its version (\code{rkeops} requires cmake >= 3.10).
}
\author{
Ghislain Durif
}
\keyword{internal}
