% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_options.R
\name{check_compile_options}
\alias{check_compile_options}
\title{Check \code{rkeops} compile options}
\usage{
check_compile_options(options)
}
\arguments{
\item{options}{a list (of class \code{rkeops_compile_options}) with the
following elements \code{rkeops_dir}, \code{build_dir}, \code{src_dir}, \code{precision},
\code{verbosity}, \code{use_cuda_if_possible}.}
}
\description{
FIXME
}
\details{
FIXME
}
\author{
Ghislain Durif
}
\keyword{internal}
