% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_check.R
\name{check_keopscore}
\alias{check_keopscore}
\title{Check if \code{keopscore} Python package is available}
\usage{
check_keopscore(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean, if TRUE (default), inform user about check result.}
}
\value{
boolean value indicating if the \code{keopscore} package is available.
}
\description{
In practice, check if \code{keopscore} Python package is installed (which is done
at \code{rkeops} package install).
}
\examples{
\dontrun{
check_keopscore()
}
}
\author{
Ghislain Durif
}
\keyword{internal}
