% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-check.R
\name{check_os}
\alias{check_os}
\title{Check OS}
\usage{
check_os(onLoad = FALSE)
}
\arguments{
\item{onLoad}{boolean indicating if the function is used when loading
the package (to avoid the warning in this case).}
}
\value{
a value 1 or 0 depending if check is ok or not.
}
\description{
This function checks the OS. Only \code{Unix} (Linux and MacOS) are supported
at the moment. Windows is not supported.
}
\details{
Return 0 if run on Windows and 1 otherwise, with a possible warning.
}
\author{
Ghislain Durif
}
\keyword{internal}
