% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_check.R
\name{check_pykeops}
\alias{check_pykeops}
\title{Check if \code{pykeops} Python package is available}
\usage{
check_pykeops(warn = TRUE)
}
\arguments{
\item{warn}{boolean, if TRUE (default), warn user about check result.}
}
\value{
boolean value indicating if the \code{pykeops} package is available.
}
\description{
In practice, check if \code{pykeops} Python package is installed (which is done
at \code{rkeops} package install).
}
\examples{
\dontrun{
check_pykeops()
}
}
\author{
Ghislain Durif
}
\keyword{internal}
