% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_check.R
\name{check_pypkg}
\alias{check_pypkg}
\title{Check if a given Python package is installed and available.}
\usage{
check_pypkg(package, warn = TRUE)
}
\arguments{
\item{package}{character string, name of package to be checked among
\code{"keopscore"} and \code{"pykeops"}.}

\item{warn}{boolean, if TRUE (default), warn user about check result.}
}
\value{
boolean value indicating if the \code{pykeops} package is available.
}
\description{
In practice, check if the given Python package is installed (which is
normally done at \code{rkeops} package install).

Should be used to test the availability of \code{keopscore} and of \code{pykeops}
packages.
}
\author{
Ghislain Durif
}
\keyword{internal}
