% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_options.R
\name{check_runtime_options}
\alias{check_runtime_options}
\title{Check \code{rkeops} runtime options}
\usage{
check_runtime_options(options)
}
\arguments{
\item{options}{a list (of class \code{rkeops_runtime_options}) with the
following \code{tagCpuGpu}, \code{tag1D2D}, \code{device_id}.}
}
\description{
FIXME
}
\details{
FIXME
}
\author{
Ghislain Durif
}
\keyword{internal}
