% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compile4float32}
\alias{compile4float32}
\title{Enable compiling of user-defined operators using float 32bits precision.}
\usage{
compile4float32()
}
\value{
None
}
\description{
Set up \code{rkeops} compile options to compile user-defined operators that use
float 32bits precision in computation.
}
\details{
\strong{Note:} Default behavior is to compile operators operators that use
float 32bits precision in computation. Hence, if you do not modify \code{rkeops}
options, you do not have to call the function \code{compile4float32} to
compile operators using float 32bits precision.

Since R only manages float 64bits or double numbers, the input and output
are casted to float 32bits before and after computations respectively.
}
\examples{
library(rkeops)
compile4float32()
}
\seealso{
\code{\link[=compile4float64]{compile4float64()}}
}
\author{
Ghislain Durif
}
