% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compile4gpu}
\alias{compile4gpu}
\title{Enable compilation of GPU-compatible user-defined operators}
\usage{
compile4gpu()
}
\description{
Set up \code{rkeops} compile options to compile user-defined operators that can
be computed on GPU.
}
\details{
Compiling GPU-compatible user-defined operators requires CUDA and \code{nvcc}
(Nvidia compiler). If not available, user-defined operators will only be
CPU-compatible.

\strong{Note:} Default behavior is to compile GPU-compatible operators thus, if
you do not modify \code{rkeops} options, it is optional to use the function
\code{compile4gpu}.
}
\examples{
\dontrun{
compile4gpu()
}
}
\seealso{
\code{\link[rkeops:use_gpu]{rkeops::use_gpu()}}
}
\author{
Ghislain Durif
}
