% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confirm_choice}
\alias{confirm_choice}
\title{Function to ask for user confirmation}
\usage{
confirm_choice(
  question = "Are you sure?",
  default_answer = TRUE,
  timeout = 20,
  max_tries = 10
)
}
\arguments{
\item{question}{character string, the question that is asked to the user.}

\item{default_answer}{logical, the default answer (if no answer is given).}

\item{max_tries}{integer, maximum number of repetition of the question in
case of bad user input.}
}
\value{
a logical \code{TRUE}/\code{FALSE} value depending on user input.
}
\description{
Return a logical \code{TRUE}/\code{FALSE} value depending on user input.
}
\details{
In non interactive mode, \code{default_answer} value is returned.

In interactive mode, the user is asked for a \code{"yes"}/\code{"no"}. If the user
provide any other answer, the question is asked again for
\code{max_tries} times.

\strong{ATTENTION}: this function may be hung indefinitely if the user does
not provide any input.
}
\author{
Ghislain Durif
}
\keyword{internal}
