% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Conj.ComplexLazyTensor}
\alias{Conj.ComplexLazyTensor}
\title{Element-wise complex conjugate.}
\usage{
\method{Conj}{ComplexLazyTensor}(z)
}
\description{
Element-wise complex conjugate.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
