% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Conj.default}
\alias{Conj.default}
\title{Complex conjugate.}
\usage{
\method{Conj}{default}()
}
\description{
Complex conjugate.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
