% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Conj}
\alias{Conj}
\title{Element-wise complex conjugate operation}
\usage{
Conj(z)
}
\arguments{
\item{z}{input for \code{\link[=Conj.default]{Conj.default()}} or \code{\link[=Conj.LazyTensor]{Conj.LazyTensor()}}.}
}
\value{
See value of \code{\link[=Conj.default]{Conj.default()}} or \code{\link[=Conj.LazyTensor]{Conj.LazyTensor()}}.
}
\description{
Symbolic element-wise complex conjugate operation for
\code{ComplexLazyTensor} objects or default element-wise complex conjugate
operation otherwise.
}
\details{
If \code{z} is a \code{LazyTensor}, see \code{\link[=Conj.LazyTensor]{Conj.LazyTensor()}}, else
see \code{\link[=Conj.default]{Conj.default()}}.
}
\examples{
# R base operation
Conj(1)
Conj(1+1i)
# LazyTensor symbolic element-wise complex conjugate
z <- matrix(1i^ (-6:5), nrow = 4)                     # complex 4x3 matrix
z_i <- LazyTensor(z, index = 'i', is_complex = TRUE)  # ComplexLazyTensor
Conj_z_i <- Conj(z_i)                                 # symbolic matrix
}
\seealso{
\code{\link[=Conj.default]{Conj.default()}}, \code{\link[=Conj.LazyTensor]{Conj.LazyTensor()}},
\code{\link[=Re]{Re()}}, \code{\link[=Im]{Im()}}, \code{\link[=Arg]{Arg()}}, \code{\link[=Mod]{Mod()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
