% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{cplx_warning}
\alias{cplx_warning}
\title{Warning for ComplexLazyTensor/LazyTensor operations.}
\usage{
cplx_warning(warn = TRUE)
}
\value{
A warning message.
}
\description{
Returns a warning message when binary operations are used with a
\code{LazyTensor} and a \code{ComplexLazyTensor}. These operations might not work
with the current rkeops version.

This function is only called in \code{real2complex.LazyTensor}, which is only
used with binary operations involving a \code{LazyTensor} and
a \code{ComplexLazyTensor}.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
