% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dllname}
\alias{create_dllname}
\title{Create name of shared library from formula and arguments}
\usage{
create_dllname(formula, args)
}
\description{
Using input formula and arguments along with current value of "precision"
option, the function \code{dllname} creates through a hash a name for the shared
library where the operators defined by the formula and arguments will be
compiled.
}
\details{
FIXME
}
\author{
Ghislain Durif
}
