% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{default.trigo.fun}
\alias{default.trigo.fun}
\alias{cos.default}
\alias{sin.default}
\alias{acos.default}
\alias{asin.default}
\alias{atan.default}
\alias{atan2.default}
\title{Trigonometric Functions}
\usage{
\method{cos}{default}(x)

\method{sin}{default}(x)

\method{acos}{default}(x)

\method{asin}{default}(x)

\method{atan}{default}(x)

\method{atan2}{default}(x, y)
}
\arguments{
\item{x, y}{numeric or complex vectors.}
}
\value{
\code{tanpi(0.5)} is \code{\link[base]{NaN}}.  Similarly for other inputs
  with fractional part \code{0.5}.
}
\description{
These functions give the obvious trigonometric functions.  They
  respectively compute the cosine, sine, tangent, arc-cosine, arc-sine,
  arc-tangent, and the two-argument arc-tangent.

  \code{cospi(x)}, \code{sinpi(x)}, and \code{tanpi(x)}, compute
  \code{cos(pi*x)}, \code{sin(pi*x)}, and \code{tan(pi*x)}.
}
\details{
The arc-tangent of two arguments \code{atan2(y, x)} returns the angle
  between the x-axis and the vector from the origin to \eqn{(x, y)},
  i.e., for positive arguments \code{atan2(y, x) == atan(y/x)}.

  Angles are in radians, not degrees, for the standard versions (i.e., a
  right angle is \eqn{\pi/2}), and in \sQuote{half-rotations} for
  \code{cospi} etc.

  \code{cospi(x)}, \code{sinpi(x)}, and \code{tanpi(x)} are accurate
  for \code{x} values which are multiples of a half.

  All except \code{atan2} are \link[base]{internal generic} \link[base]{primitive}
  functions: methods can be defined for them individually or via the
  \code{\link[base:S3groupGeneric]{Math}} group generic.
  
  These are all wrappers to system calls of the same name (with prefix
  \code{c} for complex arguments) where available.  (\code{cospi},
  \code{sinpi}, and \code{tanpi} are part of a C11 extension
  and provided by e.g.\ifelse{latex}{\out{~}}{ }{}macOS and Solaris: where not yet
  available call to \code{cos} \emph{etc} are used, with special cases
  for  multiples of a half.)
}
\examples{
x <- seq(-3, 7, by = 1/8)
tx <- cbind(x, cos(pi*x), cospi(x), sin(pi*x), sinpi(x),
               tan(pi*x), tanpi(x), deparse.level=2)
op <- options(digits = 4, width = 90) # for nice formatting
head(tx)
tx[ (x \%\% 1) \%in\% c(0, 0.5) ,]
options(op)
}
\seealso{
\link[base:Trig]{base::Trig}
}
\author{
R core team and contributors
}
