% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_options.R
\name{default_compile_options}
\alias{default_compile_options}
\title{Define a list of default options used for compilation in \code{rkeops} package}
\usage{
default_compile_options()
}
\value{
a list of class \code{rkeops_compile_options} (see
\code{\link[rkeops:compile_options]{rkeops::compile_options()}} for the detailed output).
}
\description{
To compile user-defined operators, \code{rkeops} requires compilation options.

The function \code{default_compile_options} returns a list with default values
for the corresponding options (see Details).
}
\details{
Please refer to \code{\link[rkeops:compile_options]{rkeops::compile_options()}} for a detailed description of
these options.

\strong{Note:} Default options are set up when loading \code{rkeops}. To reset
\code{rkeops} options to default, you should use the function
\code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}}. To set up a particular option, you should
use the function \code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}.
}
\examples{
\dontrun{
default_compile_options()
}
}
\seealso{
\code{\link[rkeops:compile_options]{rkeops::compile_options()}}, \code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}},
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}
}
\author{
Ghislain Durif
}
