% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_options.R
\name{default_runtime_options}
\alias{default_runtime_options}
\title{Default options used at runtime in \code{rkeops} package}
\usage{
default_runtime_options()
}
\value{
a list (of class \code{rkeops_runtime_options}) with the following
elements:
\item{build_dir}{string, path to the directory where new custom user-defined
operators will be compiled.}
\item{src_dir}{string, path to \code{keops} (C++) source files required for
compilation of user-defined operators.}
\item{precision}{string, precision for floating point computations (\code{float}
or \code{double}).}
\item{verbosity}{integer, 0-1 indicator (boolean) for verbosity.}
\item{use_cuda_if_possible}{integer, 0-1 indicator (boolean) regarding use
of GPU in computations (if possible).}
}
\description{
When using compiled user-defined operators, \code{rkeops} requires runtime
options that control how the computations are done (memory manegement,
data partition for parallelization, etc.).

The function \code{default_runtime_options} returns an object of class
\code{rkeops_runtime_options} with default values for the corresponding options
(see Details).
}
\details{
The aforementioned compile options are the following:
\itemize{
\item \code{tagCpuGpu}: \code{0} means convolution on Cpu, \code{1} means convolution on Gpu,
\code{2} means convolution on Gpu using data on device (i.e. in GPU memory).
Default value is \code{1}.
\item \code{tag1D2D}: \code{0} means 1D parallelization (over rows of matrices), and \code{1}
parallelization over blocks of rows and columns (useful with small columns
large rows matrices). Default value is \code{0}.
\item \code{device_id}: id of GPU device on the machine where the computation are
made. Default value is \code{0}. Ideally, GPU assignation should be handled
outside of \code{R} and \code{rkeops}.
}

\strong{Note:} To set up the options defined with this function, you should use
the functions \code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}} or
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}.
}
\seealso{
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}, \code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}},
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}
}
\author{
Ghislain Durif
}
