% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{-.ComplexLazyTensor}
\alias{-.ComplexLazyTensor}
\title{Subtraction or minus sign.}
\usage{
\method{-}{ComplexLazyTensor}(x, y = NA)
}
\description{
Subtraction or minus sign.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
