% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{-.LazyTensor}
\alias{-.LazyTensor}
\title{Subtraction or minus sign.}
\usage{
\method{-}{LazyTensor}(x, y = NA)
}
\description{
Subtraction or minus sign.
}
\author{
Amelie Vernay, Chloe Serre-Combe
}
\keyword{internal}
