% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{elem}
\alias{elem}
\title{Extract an element}
\usage{
elem(x, m)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{m}{an \code{integer} corresponding to the index (starting from 0) of
the element of \code{x} that will be extracted.}
}
\value{
a \code{LazyTensor} or a \code{ComplexLazyTensor}.
}
\description{
Symbolic element extraction operation for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{elem(x, m)} returns a \code{LazyTensor} that
encodes, symbolically, the \code{m+1}-th element of \code{x}, i.e. \code{x[m+1]} in
standard R notation.

\strong{IMPORTANT}: IN THIS CASE, INDICES START AT ZERO, therefore, \code{m} should
be in \verb{[0, n)}, where \code{n} is the inner dimension of \code{x}.
}
\examples{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # LazyTensor from matrix x, indexed by 'i'
m <- 2

elem_x <- elem(x_i, m)  # symbolic `m+1`-th element of `x_i`.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
