% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{exp}
\alias{exp}
\title{Element-wise exponential operation}
\usage{
exp(x)
}
\arguments{
\item{x}{input for \code{\link[=exp.default]{exp.default()}} or
\code{\link[=exp.LazyTensor]{exp.LazyTensor()}}.}
}
\value{
See value of \code{\link[=exp.default]{exp.default()}} or
\code{\link[=exp.LazyTensor]{exp.LazyTensor()}}.
}
\description{
Symbolic element-wise exponential operation for \code{LazyTensor} objects or
default element-wise exponential operation otherwise, i.e.
\eqn{e^x}.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=exp.LazyTensor]{exp.LazyTensor()}}, else
see \code{\link[=exp.default]{exp.default()}}.
}
\examples{
# R base operation
exp(0)
exp(1:10)
# LazyTensor symbolic element-wise exponential
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Exp_x <- exp(x_i)                   # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=exp.default]{exp.default()}}, \code{\link[=exp.LazyTensor]{exp.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
