% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{exp1j.LazyTensor}
\alias{exp1j.LazyTensor}
\title{Element-wise "complex exponential of 1j x" operation.}
\usage{
\method{exp1j}{LazyTensor}(x)
}
\description{
Element-wise "complex exponential of 1j x" operation.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
