% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-format_var_aliase.R
\name{format_var_aliases}
\alias{format_var_aliases}
\title{Format formula argument list in triplet (type, dimension, position)}
\usage{
format_var_aliases(args)
}
\description{
The function \code{format_var_aliases} formats KeOps formula arguments to be
understood by the C++ code.
}
\details{
\code{e^(lambda*||x_i - y_j||^2) * beta_j} where \code{x_i}, \code{y_j}
and \code{beta_j} are 3d vectors\preformatted{formula = "Sum_Reduction(Exp(lambda * SqNorm2(x-y)) * beta, 0)"
args = c("x=Vi(3)", "y=Vj(3)", 
         "beta=Vj(3)", "lambda=Pm(1)")
}

|---------|-------------------------|-----------|
| keyword | meaning                 | type      |
|---------|-------------------------|-----------|
| \code{Vi}    | variable indexed by \code{i} | \code{0}       |
| \code{Vj}    | variable indexed by \code{j} | \code{1}       |
| \code{Pm}    | parameter               | \code{2}       |
|---------|-------------------------|-----------|
}
\keyword{internal}
