% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_build_dir}
\alias{get_build_dir}
\title{Getter for package build directory}
\usage{
get_build_dir(pkg = "rkeops", create = TRUE)
}
\arguments{
\item{pkg}{name (string) of the R package, default is "rkeops".}

\item{create}{boolean indicating if the corresponding directory
should be created if missing or not. Default value is TRUE.}
}
\value{
path to the corresponding directory.
}
\description{
Return the path to the directory \code{build} where \code{keops} custom
operators will be compiled.
}
\details{
This path is generaly something like
\verb{/path/to/R_package_install/rkeops/build}. The correspondinging
directory can be created if not existing.
\strong{Note:} when running tests in the development project \code{keops} without
installing \code{rkeops}, for consistency reason, the function returns
\verb{/path/to/keops/rkeops/inst/build}.
}
\author{
Ghislain Durif
}
\keyword{internal}
