% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-get_cmake.R
\name{get_cmake}
\alias{get_cmake}
\title{Return path to cmake executable}
\usage{
get_cmake(path = NULL)
}
\arguments{
\item{path}{text string (optional), path where to search
\code{cmake} (e.g. \verb{/path/to/dir} on Unix system). Default value
is NULL and the function searches for \code{cmake} in the
standard PATH (on unix system).}
}
\value{
path to cmake executable.
}
\description{
This function returns the path to the cmake executable.
}
\details{
If no path is supplied, the function searches for the
\code{cmake} command in the standard PATH (on Unix system).
If a path is supplied (e.g. \verb{/path/to/dir} on Unix
system), the function search for the \code{cmake} command
in this path.

\strong{Note}: for the moment, this does not work on
Windows OS.
}
\author{
Ghislain Durif
}
\keyword{internal}
