% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keops_grad.R
\name{get_gradient_formula}
\alias{get_gradient_formula}
\title{Generate formula and input args list for the Gradient of an existing formula}
\usage{
get_gradient_formula(formula, args, var, args_info)
}
\arguments{
\item{formula}{text string, an operator formula (see Details).}

\item{args}{vector of text string, formula arguments (see Details).}

\item{var}{a text string or an integer number indicating regarding to which
variable/parameter (given by name or by position starting at 0) the
gradient of the formula should be computed.}

\item{args_info}{list, internal (output of
\code{rkeops:::parse_args(formula,args)})}
}
\value{
a list with the following element:
\itemize{
\item \code{new_formula}: character string corresponding to the Gradient formula.
\item \code{new_args}: corresponding enriched arguments (see \code{\link[=keops_grad]{keops_grad()}}
for more details).
}
}
\description{
Generate formula and input args list for the Gradient of an existing formula
}
\author{
Ghislain Durif
}
\keyword{internal}
