% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Find which OS is running}
\usage{
get_os()
}
\value{
a character string containing the OS name.
}
\description{
Return the name of the currently running OS.
}
\details{
Possible output among \code{"linux"}, \code{"macos"}, \code{"windows"}.
}
\author{
Ghislain Durif
}
\keyword{internal}
