% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_dir}
\alias{get_pkg_dir}
\title{Getter for package installation directory}
\usage{
get_pkg_dir(pkg = "rkeops")
}
\arguments{
\item{pkg}{name (string) of package, default is "rkeops".}
}
\value{
path to the package local installation.
}
\description{
Return the path to the directory where the package
provided in input is installed on the system.
}
\author{
Ghislain Durif
}
\keyword{internal}
