% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keops_utils.R
\name{get_rkeops_build_dir}
\alias{get_rkeops_build_dir}
\title{Get RKeOps build directory}
\usage{
get_rkeops_build_dir()
}
\value{
character string, path to RKeOps build directory.
}
\description{
Get the path to the folder where all dynamic library files generated
from compilations of user-defined operators are defined.
}
\details{
When compiling a user-defined operators, a shared object (\code{.so}) library
(or dynamic link library, \code{.dll}) file is created in RKeOps build
directory (located in the \code{.cache} folder in your home). For every
new operators, such a file is created.

Calling \code{get_rkeops_build_dir()} gives the location of this specific
directory on your system.

You can use \code{\link[=ls_rkeops_build_dir]{ls_rkeops_build_dir()}} to list RKeOps build
directory content, and you can use \code{\link[=clean_rkeops]{clean_rkeops()}} to delete its
content.
}
\examples{
\dontrun{
get_rkeops_build_dir()
}
}
\seealso{
\code{\link[=ls_rkeops_build_dir]{ls_rkeops_build_dir()}}, \code{\link[=clean_rkeops]{clean_rkeops()}}
}
\author{
Ghislain Durif
}
