% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_options.R
\name{get_rkeops_cache_dir}
\alias{get_rkeops_cache_dir}
\title{Get RKeOps cache directory}
\usage{
get_rkeops_cache_dir()
}
\value{
character string, path to RKeOps cache directory.
}
\description{
Get the path to the cache folder where rkeops operator
compilation byproducts will be stored to be re-used for further use
(and avoid unnecessary recompilation).
}
\details{
When compiling a user-defined operators or running a LazyTensor-based
operations, a shared object (\code{.so}) library (or dynamic link library,
\code{.dll}) file is created in RKeOps cache directory.

For every new operators or new LazyTensor operations, such a file is
created, to be reused next time the same computations is done (then saving
compilation time).

Calling \code{get_rkeops_cache_dir()} gives the location of this specific
directory on your system.

You can use \code{\link[=stat_rkeops_cache_dir]{stat_rkeops_cache_dir()}} to verify rkeops cache
directory disk usage, and you can use \code{\link[=clean_rkeops]{clean_rkeops()}}
to (fully or partially) delete its content.

\strong{Note:} see \code{\link[=default_rkeops_cache_dir]{default_rkeops_cache_dir()}} for more details about
the default rkeops cache directory.
}
\examples{
get_rkeops_cache_dir()
}
\seealso{
\code{\link[=set_rkeops_cache_dir]{set_rkeops_cache_dir()}}, \code{\link[=stat_rkeops_cache_dir]{stat_rkeops_cache_dir()}},
\code{\link[=clean_rkeops]{clean_rkeops()}}
}
\author{
Ghislain Durif
}
