% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-get_options.R
\name{get_rkeops_option}
\alias{get_rkeops_option}
\title{Get the current value of a specific compile or runtime options of \code{rkeops}}
\usage{
get_rkeops_option(option)
}
\arguments{
\item{option}{string, name of the options to set up (see Details).}
}
\value{
the value of the requested option (see Details).
}
\description{
The function \code{get_rkeops_option} returns the current value of a specific
\code{rkeops} option (in \code{R} global options scope) provided as input.
}
\details{
\code{rkeops} uses two sets of options: compile options \code{rkeops_dir},
\code{build_dir}, \code{src_dir}, \code{precision}, \code{verbosity}, \code{use_cuda_if_possible}
(see \code{\link[rkeops:compile_options]{rkeops::compile_options()}}), and runtime options \code{tagCpuGpu},
\code{tag1D2D}, \code{device_id} (see \code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}).

These options define the behavior of \code{rkeops} when compiling or when
running new user-defined operators.

With the function \code{get_rkeops_option}, you get the value of a specific
\code{rkeops} options among \code{rkeops_dir}, \code{build_dir}, \code{src_dir}, \code{precision},
\code{verbosity}, \code{use_cuda_if_possible} or \code{tagCpuGpu}, \code{tag1D2D}, \code{device_id}.

These options are set with the functions \code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}} and
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}. To know which values are allowed for which
options, you can check \code{\link[rkeops:compile_options]{rkeops::compile_options()}} and
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}.
}
\seealso{
\code{\link[rkeops:get_rkeops_options]{rkeops::get_rkeops_options()}}, \code{\link[rkeops:compile_options]{rkeops::compile_options()}},
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}, \code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}},
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}
}
\author{
Ghislain Durif
}
