% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{\%*\%}
\alias{\%*\%}
\title{Matrix multiplication.}
\usage{
x %*% y
}
\arguments{
\item{x}{A LazyTensor, a vector or a matrix of numeric values, or a scalar value.}

\item{y}{A LazyTensor, a vector or a matrix of numeric values, or a scalar value.}
}
\value{
A matrix
}
\description{
Symbolic binary operation for element-wise matrix multiplication operator.
}
\details{
If \code{x} or \code{y} is a \code{LazyTensor}, \code{x \%*\% y} .... TODO (sum_reduction)
If none of the arguments is a \code{LazyTensor}, is equivalent to the "\%*\%" R operator.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
y <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'
y_j <- LazyTensor(x, index = 'j')   # creating LazyTensor from matrix x, indexed by 'j'
x_mult_y <- x_i \%*\% y_j             
}
}
\author{
Ghislain Durif
}
