% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{ifelse.default}
\alias{ifelse.default}
\title{if-else function.}
\usage{
\method{ifelse}{default}(x, a, b)
}
\description{
if-else function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
