% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Im.ComplexLazyTensor}
\alias{Im.ComplexLazyTensor}
\title{Element-wise imaginary part of complex.}
\usage{
\method{Im}{ComplexLazyTensor}(z)
}
\description{
Element-wise imaginary part of complex.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
