% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Im.default}
\alias{Im.default}
\title{Imaginary part of complex.}
\usage{
\method{Im}{default}()
}
\description{
Imaginary part of complex.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
