% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{imag2complex.LazyTensor}
\alias{imag2complex.LazyTensor}
\title{Element-wise "imag 2 complex" operation.}
\usage{
\method{imag2complex}{LazyTensor}(x)
}
\description{
Element-wise "imag 2 complex" operation.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
