% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{imaginary2complex}
\alias{imaginary2complex}
\alias{imag2complex}
\alias{imag2complex.LazyTensor}
\alias{imag2complex.ComplexLazyTensor}
\title{Element-wise "imaginary to complex" operation.}
\usage{
imag2complex(x)

\method{imag2complex}{LazyTensor}(x)

\method{imag2complex}{ComplexLazyTensor}(x)
}
\arguments{
\item{x}{a \code{LazyTensor}.}
}
\value{
An object of class \code{ComplexLazyTensor}.
}
\description{
Symbolic operation for element-wise "imaginary to complex".
}
\details{
\code{imag2complex(x)} returns a \code{ComplexLazyTensor} that encodes,
symbolically, the element-wise "imaginary to complex" transformation
of \code{x} (i.e. with additional zero real part: \code{0 + x*i}).

\strong{Note}: \code{imag2complex(x)} will exit with an error if \code{x} is
a \code{ComplexLazyTensor}.
}
\examples{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, "i")           # creating LazyTensor from matrix x,
                                    # indexed by 'i'
z <- imag2complex(x_i)              # ComplexLazyTensor object
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
