% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_rkeops}
\alias{install_rkeops}
\title{Install RKeOps requirements}
\usage{
install_rkeops(
  envname = NULL,
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  python_version = NULL,
  pip = FALSE,
  ...,
  pip_ignore_installed = ignore_installed,
  ignore_installed = FALSE
)
}
\arguments{
\item{envname}{The name, or full path, of the environment in which Python
packages are to be installed. When \code{NULL} (the default), the active
environment as set by the \code{RETICULATE_PYTHON_ENV} variable will be used;
if that is unset, then the \code{r-reticulate} environment will be used.}

\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{python_version}{The requested Python version. Ignored when attempting
to install with a Python virtual environment.}

\item{pip}{Boolean; use \code{pip} for package installation? This is only relevant
when Conda environments are used, as otherwise packages will be installed
from the Conda repositories.}

\item{...}{Additional arguments passed to \code{\link[reticulate:conda_install]{conda_install()}}
or \code{\link[reticulate:virtualenv_install]{virtualenv_install()}}.}

\item{pip_ignore_installed, ignore_installed}{Boolean; whether pip should
ignore previously installed versions of the requested packages. Setting
this to \code{TRUE} causes pip to install the latest versions of all
dependencies into the requested environment. This ensure that no
dependencies are satisfied by a package that exists either in the site
library or was previously installed from a different--potentially
incompatible--distribution channel. (\code{ignore_installed} is an alias for
\code{pip_ignore_installed}, \code{pip_ignore_installed} takes precedence).}
}
\description{
Install requirements (i.e. Python packages) for RKeOps to work.
}
\details{
RKeOps now uses PyKeOps Python package under the hood thanks to the
\href{https://rstudio.github.io/reticulate/}{\code{reticulate}} R package that
provides an "R Interface to Python".

We recommend to use a dedicated Python environment (through \code{reticulate})
to install and use RKeOps Python dependencies.

\code{install_rkeops()} is just a wrapper around \code{\link[reticulate:py_install]{reticulate::py_install()}}
and takes the same input arguments.

See the specific \strong{"Using RKeOps" \href{https://www.kernel-operations.io/rkeops/articles/using_rkeops.html}{article}}
or the corresponding vignette: \code{vignette("using_rkeops", package = "rkeops")}

\strong{Important:} Python is a requirement as an intern machinery for the
package to work but you will not need to create nor manipulate Python
codes to use the RKeOps package.
}
\examples{
\dontrun{
## setup Python virtual environment
# create Python virtualenv
reticulate::virtualenv_create("rkeops")
# activate python environment
reticulate::use_virtualenv(virtualenv = "rkeops", required = TRUE)

## install rkeops requirements in Python virtual environment
install_rkeops()

## check install
check_rkeops()
}
}
\author{
Ghislain Durif
}
