% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{is.LazyScalar}
\alias{is.LazyScalar}
\title{is.LazyScalar?}
\usage{
is.LazyScalar(x)
}
\arguments{
\item{x}{An object we want to know if it is a \code{LazyScalar}.}
}
\value{
A boolean, TRUE or FALSE.
}
\description{
Checks whether if the given input is a \code{LazyTensor}encoding
a single scalar value. That is, if the input is a fixed parameter
\code{LazyTensor} of dimension 1.
}
\details{
If \code{x} is a fixed parameter \code{LazyTensor} with dimension 1,
\code{is.LazyScalar(x)} returns TRUE, else, returns FALSE.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
