% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_installed}
\alias{is_installed}
\title{Check if \code{rkeops} is installed on the system}
\usage{
is_installed()
}
\value{
1 if ok, 0 otherwise
}
\description{
When running automatic tests without building and installing the package,
i.e. with \code{devtools::test()}, the package file structure is different
from the structure of a built and installed package (obtained with
\code{devtools::check()}). For instance, the directory \code{inst/include} is
replaced by \code{include}, or \code{src} by \code{lib}. Check functions inside \code{rkeops}
have different behavior between these two cases.

The function \code{is_installed} checks the availability of the compiled
function \code{\link[=is_compiled]{is_compiled()}}. If available, it means that \code{rkeops} is
installed on the system.
}
\seealso{
\code{\link[=is_installed]{is_installed()}}
}
\author{
Ghislain Durif
}
\keyword{internal}
