% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Kmin_reduction}
\alias{Kmin_reduction}
\title{Kmin reduction.}
\usage{
Kmin_reduction(x, K, index)
}
\arguments{
\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{K}{An \code{integer} corresponding to the  number of minimal values required.}

\item{index}{A \code{character} corresponding to the reduction dimension that should
be either \strong{i} or \strong{j} to specify whether if the reduction is indexed by
\strong{i} (rows) or \strong{j} (columns).}
}
\value{
A matrix corresponding to the Kmin reduction.
}
\description{
K-Min reduction. Redirects to \code{Kmin} function.
}
\details{
If \code{x} is a \code{LazyTensor} or a \code{ComplexLazyTensor},
\code{Kmin(x, K, index)} will:
\itemize{
\item{if \strong{index = i},}{ return the \strong{K} minimal values of \strong{x}
over the \strong{i} indices (rows);}
\item{if \strong{index = j},}{ return the \strong{K} minimal values of \strong{x}
over the \strong{j} indices (columns).}
}
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # LazyTensor from matrix x, indexed by 'i'
K <- 2
kmin_red_x <- Kmin_reduction(x_i, K, "i")   # Kmin reduction, indexed by "i"
}
}
\seealso{
\code{\link[=Kmin]{Kmin()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
