% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{log.LazyTensor}
\alias{log.LazyTensor}
\title{Element-wise natural logarithm operation}
\usage{
\special{log(x)}
}
\arguments{
\item{x}{a \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values,
or a scalar value.}

\item{base}{not used, only present for method compatibility with
corresponding generic function.}
}
\value{
An object of class \code{LazyTensor}.
}
\description{
Symbolic element-wise natural logarithm operation for \code{LazyTensor}
objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{log(x)} returns a \code{LazyTensor} that
encodes, symbolically, the element-wise natural logarithm of \code{x}, i.e.
\eqn{\log{x}}.
}
\examples{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Log_x <- log(x_i)                   # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=log]{log()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
