% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{log}
\alias{log}
\title{Element-wise natural logarithm operation}
\usage{
log(x, base = exp(1))
}
\arguments{
\item{x}{input for \code{\link[=log.default]{log.default()}} or
\code{\link[=log.LazyTensor]{log.LazyTensor()}}.}

\item{base}{numeric, see \code{\link[=log.default]{log.default()}}, not used with \code{LazyTensor}
input.}
}
\value{
See value of \code{\link[=log.default]{log.default()}} or
\code{\link[=log.LazyTensor]{log.LazyTensor()}}.
}
\description{
Symbolic element-wise natural logarithm operation for \code{LazyTensor}
objects or default element-wise natural logarithm operation otherwise,
i.e. \eqn{\log(x)}.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=log.LazyTensor]{log.LazyTensor()}}, else
see \code{\link[=log.default]{log.default()}}.
}
\examples{
# R base operation
log(1)
log(1:10)
# LazyTensor symbolic element-wise natural logarithm
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Log_x <- log(x_i)                   # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=log.default]{log.default()}}, \code{\link[=log.LazyTensor]{log.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
