% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{logical.or.default}
\alias{logical.or.default}
\alias{|.default}
\title{Logical "or"}
\usage{
\method{|}{default}(x, y)
}
\description{
Logical "or"
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
