% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{logical}
\alias{logical}
\alias{|.default}
\title{Logical.or}
\usage{
\method{|}{default}(x, y)
}
\arguments{
\item{x, y}{\code{\link[base]{raw}}, \code{\link[base]{logical}} or \sQuote{number-like} vectors (i.e., of
    types \code{\link[base]{double}} (class \code{\link[base]{numeric}}),
    \code{\link[base]{integer}} and \code{\link[base]{complex}}), or objects for
    which methods have been written.}
}
\value{
For \code{!}, a logical or raw vector(for raw \code{x}) of the same
  length as \code{x}: names, dims and dimnames are copied from \code{x},
  and all other attributes (including class) if no coercion is done.

  For \code{|}, \code{&} and \code{xor} a logical or raw vector.  If
  involving a zero-length vector the result has length zero.  Otherwise,
  the elements of shorter vectors are recycled as necessary (with a
  \code{\link[base]{warning}} when they are recycled only \emph{fractionally}).
  The rules for determining the attributes of the result are rather
  complicated.  Most attributes are taken from the longer argument, the
  first if they are of the same length.  Names will be copied from the
  first if it is the same length as the answer, otherwise from the
  second if that is.  For time series, these operations are allowed only
  if the series are compatible, when the class and \code{\link{tsp}}
  attribute of whichever is a time series (the same, if both are) are
  used.  For arrays (and an array result) the dimensions and dimnames
  are taken from first argument if it is an array, otherwise the second.

  For \code{||}, \code{&&} and \code{isTRUE}, a length-one logical vector.
}
\description{
Logical "or" operator.
}
\details{
See \link[base:Logic]{base::Logic} for more details.
}
\examples{
TRUE | FALSE
x <- 1:10
(x < 2) | (x > 8)
}
\seealso{
\link[base:Logic]{base::Logic}
}
\author{
R core team and contributors
}
