% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{logsumexp}
\alias{logsumexp}
\alias{logsumexp_reduction}
\title{Log-Sum-Exp reduction.}
\usage{
logsumexp(x, index, weight = NULL)

logsumexp_reduction(x, index, weight = NULL)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{a \code{character} corresponding to the reduction dimension that
should be either \code{"i"} or \code{"j"} to specify whether if the reduction is
indexed by \code{"i"} or \code{"j"}.}

\item{weight}{an optional object (\code{LazyTensor} or \code{ComplexLazyTensor}) that
specifies scalar or vector-valued weights. \code{NULL} by default and not used.
\strong{Not supported at the moment}.}
}
\value{
a matrix corresponding to the Log-Sum-Exp reduction.
}
\description{
Log-Sum-Exp reduction for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor} or a \code{ComplexLazyTensor},
\code{logsumexp(x, index, weight)} will:
\itemize{
\item if \code{index = "i"}, return the Log-Sum-Exp reduction of \code{x}
over the \code{i} indexes;
\item if \code{index = "j"}, return the Log-Sum-Exp reduction of \code{x}
over the \code{j} indexes.
}

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Reductions"}.
}
\examples{
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) 
x_i <- LazyTensor(x, index = 'i') 
y <- matrix(runif(100 * 3), 100, 3)
y_j <- LazyTensor(y, index = 'j')
w <- matrix(runif(100 * 3), 100, 3) # weight LazyTensor
w_j <- LazyTensor(w, index = 'j')

S_ij = sum((x_i - y_j)^2)
                                         
logsumexp_x <- logsumexp(S_ij, 'i')      # logsumexp reduction
                                         # over the 'i' indices

\dontrun{
# note: expected error, not supported at the moment
logsumexp_xw <- logsumexp(S_ij, 'i', w_j) # logsumexp reduction with weights
                                          # over the 'i' indices
}
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) 
x_i <- LazyTensor(x, index = 'i') 
y <- matrix(runif(100 * 3), 100, 3)
y_j <- LazyTensor(y, index = 'j')
w <- matrix(runif(150 * 3), 150, 3) # weight LazyTensor
w_j <- LazyTensor(y, index = 'j')

S_ij = sum( (x_i - y_j)^2 )
                                         
logsumexp_x <- logsumexp_reduction(S_ij, 'i')  # logsumexp reduction without
                                               # weight over the 'i' indices

\dontrun{
# note: expected error, not supported at the moment
logsumexp_xw <- logsumexp_reduction(S_ij, 'i', w_j) # logsumexp reduction 
                                                    # over the 'i' indices
}
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
