% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keops_utils.R
\name{ls_rkeops_build_dir}
\alias{ls_rkeops_build_dir}
\title{List RKeOps build directory content}
\usage{
ls_rkeops_build_dir()
}
\value{
output of \code{\link[=file.info]{file.info()}} function.
}
\description{
List all files in the folder where all dynamic library files generated
from compilations of user-defined operators are defined.
}
\details{
When compiling a user-defined operators, a shared object (\code{.so}) library
(or dynamic link library, \code{.dll}) file is created in RKeOps build
directory (located in the \code{.cache} folder in your home). For every
new operators, such a file is created.

Calling \code{ls_rkeops_build_dir()} lists all files in RKeOps build
directory.

You can use \code{\link[=get_rkeops_build_dir]{get_rkeops_build_dir()}} get the path to RKeOps build
directory, and you can use \code{\link[=clean_rkeops]{clean_rkeops()}} to delete its content.
}
\examples{
\dontrun{
ls_rkeops_build_dir()
}
}
\seealso{
\code{\link[=get_rkeops_build_dir]{get_rkeops_build_dir()}}, \code{\link[=clean_rkeops]{clean_rkeops()}}
}
\author{
Ghislain Durif
}
