% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{matmult.default}
\alias{matmult.default}
\alias{\%*\%.default}
\title{Matrix multiplication}
\usage{
\method{\%*\%}{default}(x, y)
}
\arguments{
\item{x, y}{numeric or complex matrices or vectors.}
}
\value{
A double or complex matrix product.  Use \code{\link[base]{drop}} to remove
  dimensions which have only one level.
}
\description{
Multiplies two matrices, if they are conformable.  If one argument is
  a vector, it will be promoted to either a row or column matrix to make
  the two arguments conformable.  If both are vectors of the same
  length, it will return the inner product (as a matrix).
}
\details{
When a vector is promoted to a matrix, its names are not
  promoted to row or column names, unlike \code{\link[base]{as.matrix}}.

  Promotion of a vector to a 1-row or 1-column matrix happens when one
  of the two choices allows \code{x} and \code{y} to get conformable
  dimensions.

  This operator is a generic function: methods can be written for it
  individually or via the \code{\link[base:S3groupGeneric]{matOps}} group
  generic function; it dispatches to S3 and S4 methods.  Methods need to be
  written for a function that takes two arguments named \code{x} and \code{y}.
}
\examples{
x <- 1:4
(z <- x \%*\% x)    # scalar ("inner") product (1 x 1 matrix)
drop(z)             # as scalar

y <- diag(x)
z <- matrix(1:12, ncol = 3, nrow = 4)
y \%*\% z
y \%*\% x
x \%*\% z
}
\seealso{
\link[base:matmult]{base::matmult}
}
\author{
R core team and contributors
}
