% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{max_argmax}
\alias{max_argmax}
\alias{max_argmax_reduction}
\title{Max-ArgMax reduction}
\usage{
max_argmax(x, index)

max_argmax_reduction(x, index)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{a \code{character} corresponding to the wanted reduction dimension,
either \code{"i"} or \code{"j"}, to specify whether if the Max-ArgMax reduction is done
along the index \code{i} or \code{j}.}
}
\value{
an array storing the result of the specified Max-ArgMax
reduction otherwise.
}
\description{
Max-ArgMax reduction for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{max_argmax(x, index)} will :
\itemize{
\item if \code{index = "i"}, return the Max-ArgMax reduction of \code{x} over the \code{i}
indexes.
\item if \code{index = "j"}, return the Max-ArgMax reduction of \code{x} over the \code{j}
indexes.
}

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Reductions"}.
}
\examples{
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

max_argmax_x <- max_argmax(x_i, "i")  # max argmax reduction indexed by 'i'
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

max_argmax_red <- max_argmax_reduction(x_i, "i")  # max argmax reduction 
                                                  # indexed by 'i'
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
