% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{min}
\alias{min}
\title{Minimum operation or minimum reduction}
\usage{
min(x, ...)
}
\arguments{
\item{x}{input for \code{\link[=min.default]{min.default()}} or \code{\link[=min.LazyTensor]{min.LazyTensor()}}.}

\item{...}{optional arguments for \code{\link[=min.default]{min.default()}} or
\code{\link[=min.LazyTensor]{min.LazyTensor()}}.}
}
\value{
Minimum of input.
}
\description{
Minimum operation for vector/matrix/array or \code{LazyTensor}, or minimum
reduction for \code{LazyTensor}.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=min.LazyTensor]{min.LazyTensor()}}, else see
\code{\link[=min.default]{min.default()}}.
}
\examples{
# R base operation
min(1:10)
min(c(NA, 1, 2), na.rm = TRUE)
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
# LazyTensor operation
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

min_x <- min(x_i)                   # LazyTensor object
min_red_x <- min(x_i, "i")          # min reduction indexed by 'i'
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
