% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{mod}
\alias{mod}
\title{Element-wise modulo with offset function}
\usage{
mod(x, ...)
}
\arguments{
\item{x}{input for \code{\link[=ifelse.default]{ifelse.default()}} or any other \code{mod} function
(if existing).}

\item{...}{optional additional input arguments.}
}
\value{
See value of \code{\link[=mod.LazyTensor]{mod.LazyTensor()}} or any other \code{mod} function
(if existing).
}
\description{
Symbolic element-wise modulo with offset function  (ternary operation) for
\code{LazyTensor} objects  or any other \code{mod} function (if existing).
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=ifelse.LazyTensor]{ifelse.LazyTensor()}}, else
see the corresponding \code{mod} function (if existing) help page.

\strong{Warning}: Do not confuse with \code{\link[=Mod]{Mod()}}.
}
\examples{
# basic example
D <- 3
M <- 100
N <- 150
P <- 200
x <- matrix(runif(M * D), M, D)
y <- matrix(runif(N * D), N, D)
z <- matrix(runif(P * D), P, D)
x_i <- LazyTensor(x, index = 'i')
y_j <- LazyTensor(y, index = 'j')
z_i <- LazyTensor(z, index = 'i')

# call mod function
mod_x72 <- mod(x_i, 7, 2)

# works also with LazyTensors with same inner dimension or dimension 1
mod_xyz <- mod(x_i, y_j, z_i)
}
\seealso{
\code{\link[=mod.LazyTensor]{mod.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
