% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_or_error}
\alias{msg_or_error}
\title{Print message if used on startup or raise error otherwise}
\usage{
msg_or_error(msg, onLoad = FALSE)
}
\arguments{
\item{msg}{character string, text message.}

\item{onLoad}{boolean, indicating whether the function is called from
startup or not.}
}
\description{
Different behavior for message raised by checks.
}
\details{
If \code{onLoad=TRUE}, print the message \code{msg}. If \code{onLoad=FALSE}, raise on error
with message \code{msg}.
}
\author{
Ghislain Durif
}
\keyword{internal}
