% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{|.LazyTensor}
\alias{|.LazyTensor}
\title{Euclidean scalar product}
\usage{
\method{|}{LazyTensor}(x, y)
}
\description{
Euclidean scalar product
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
