% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-format_var_aliase.R
\name{parse_extra_args}
\alias{parse_extra_args}
\title{Parse formula for extra arguments not encoded in argument aliases}
\usage{
parse_extra_args(formula, args)
}
\arguments{
\item{formula}{text string, an operator formula (see Details).}

\item{args}{vector of text string, formula arguments (see Details).}
}
\value{
FIXME
}
\description{
FIXME
}
\details{
Parse the formula for string such as
\itemize{
\item \code{"YY(<pos>,<dim>)"} where \code{YY} can be a formula input argument type
(\code{Vi}, \code{Vj} or \code{Pm}), \code{<pos>} is the position of the corresponding input
argument, and \code{<dim>} its inner dimension.
\item \code{"Var(<pos>,<dim>,<type>)"} where \code{<pos>} and \code{<dim>} are the position and
inner dimension (c.f. previous point) and \code{<type>} is an integer encoding
the formula input argument type with the following relation:
}

|---------|-------------------------|-----------|
| keyword | meaning                 | type      |
|---------|-------------------------|-----------|
| \code{Vi}    | variable indexed by \code{i} | \code{0}       |
| \code{Vj}    | variable indexed by \code{j} | \code{1}       |
| \code{Pm}    | parameter               | \code{2}       |
|---------|-------------------------|-----------|
}
\author{
Ghislain Durif
}
\keyword{internal}
