% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{preprocess_reduction}
\alias{preprocess_reduction}
\title{Reduction preprocess.}
\usage{
preprocess_reduction(x, opstr, index, opt_arg = NA)
}
\description{
Reduction preprocess.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
