% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pykeops}
\alias{pykeops}
\title{Glbal reference to main PyKeOps module}
\format{
An object of class \code{python.builtin.module}
}
\value{
the \code{pykeops} Python module
}
\description{
The \code{pykeops} module object is an internal reference to the PyKeOps
Python package that is used under the hood by \code{rkeops}.
}
\keyword{internal}
