% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{R_makeconf_path}
\alias{R_makeconf_path}
\title{Get path to R default Makeconf file}
\usage{
R_makeconf_path(path = NULL)
}
\arguments{
\item{path}{test string, path to custom Makeconf file. Default value is
NULL and R default Makeconf file is used.}
}
\value{
path to Makeconf file.
}
\description{
Default R compile and linking options are defined in the Makeconf files.
The function return the path to this file.
}
\details{
The default compile and linking options are required to compile new user-
defined operators. We use the option defined in the R default Makeconf file
(in particular used by R CMD SHLIB).
}
\author{
Ghislain Durif
}
\keyword{internal}
