% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_varname}
\alias{random_varname}
\title{Helper function to generate random variable name for gradient input in
formula}
\usage{
random_varname(prefix = "", len = 5)
}
\arguments{
\item{prefix}{character string, prefix to add to the variable name.}

\item{len}{integer, number of random character in the variable name.}
}
\value{
character string, name of the python environment.
}
\description{
Helper function to generate random variable name for gradient input in
formula
}
\keyword{internal}
