% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Re.default}
\alias{Re.default}
\title{Real part of complex.}
\usage{
\method{Re}{default}()
}
\description{
Real part of complex.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
