% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{real2complex.ComplexLazyTensor}
\alias{real2complex.ComplexLazyTensor}
\title{Element-wise "real 2 complex" operation.}
\usage{
\method{real2complex}{ComplexLazyTensor}(x)
}
\description{
Element-wise "real 2 complex" operation.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
