% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{reduction.LazyTensor}
\alias{reduction.LazyTensor}
\title{Reduction operation}
\usage{
reduction.LazyTensor(x, opstr, index, opt_arg = NULL)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{opstr}{a \code{string} formula (like "Sum" or "Max").}

\item{index}{a \code{character} corresponding to the wanted reduction dimension,
either \code{"i"} or \code{"j"}, to specify whether if the reduction is done along
the index \code{i} or \code{j}.}

\item{opt_arg}{an optional argument: an \code{integer} (e.g. for \code{Kmin}
reduction), a \code{character}, a \code{LazyTensor} or a \code{ComplexLazyTensor}. \code{NULL}
if not used (default).}
}
\value{
an array storing the result of the specified reduction.
}
\description{
Applies a reduction to a \code{LazyTensor}.
}
\details{
\code{reduction.LazyTensor(x, opstr, index)} will :
\itemize{
\item if \code{index = "i"}, return the \code{opstr} reduction of \code{x} over the "i" indexes;
\item if \code{index = "j"}, return the \code{opstr} reduction of \code{x} over the "j" indexes.
}
}
\examples{
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

red_x <- reduction.LazyTensor(x_i, "Sum", "i")
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
