% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{relu.default}
\alias{relu.default}
\title{Element-wise ReLU function}
\usage{
\method{relu}{default}(x)
}
\arguments{
\item{x}{numeric vectors or objects which can be coerced to such.}
}
\value{
Vector or array containing the element-wise ReLU of \code{x}.
}
\description{
Element-wise ReLU (Rectified Linear Unit) function on numeric vectors
(or objects which can be coerced to them).
}
\details{
The ReLU function \insertCite{fukushima_cognitron_1975}{rkeops} is defined
as follows: \code{relu(x)} returns \code{0} if \code{x < 0}, \code{x} if \code{x >= 0}.
}
\examples{
relu(4)
relu(-10:10)
}
\references{
\insertRef{fukushima_cognitron_1975}{rkeops}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
