% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{relu.LazyTensor}
\alias{relu.LazyTensor}
\title{Element-wise ReLU function}
\usage{
\method{relu}{LazyTensor}(x)
}
\arguments{
\item{x}{a \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values,
or a scalar value.}
}
\value{
An object of class \code{LazyTensor}.
}
\description{
Symbolic element-wise ReLU (Rectified Linear Unit) function for
\code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{relu(x)} returns a \code{LazyTensor} that
encodes, symbolically, the element-wise ReLU of \code{x}.

The ReLU function \insertCite{fukushima_cognitron_1975}{rkeops} is defined
as follows: \code{relu(x)} returns \code{0} if \code{x < 0}, \code{x} if \code{x >= 0}.
}
\examples{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
ReLU_x <- relu(x_i)                 # symbolic matrix, 150 rows and 3 columns
}
\references{
\insertRef{fukushima_cognitron_1975}{rkeops}
}
\seealso{
\code{\link[=relu]{relu()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
