% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{rsqrt.default}
\alias{rsqrt.default}
\title{Inverse square root.}
\usage{
\method{rsqrt}{default}(x)
}
\description{
Inverse square root.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
