% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_options.R
\name{runtime_options}
\alias{runtime_options}
\title{User-defined options used at runtime in \code{rkeops} package}
\usage{
runtime_options(tagCpuGpu = 1, tag1D2D = 0, device_id = 0)
}
\arguments{
\item{tagCpuGpu}{integer, indicator for CPU or GPU computations (see
Details). Default value is \code{1}.}

\item{tag1D2D}{integer, indicator regarding data partitioning for
parallelization (see Details). Default value is \code{0}.}

\item{device_id}{integer, id of GPU device on the machine (see Details).
Default value is \code{0}.}
}
\value{
a list (of class \code{rkeops_runtime_options}) with the following
elements:
\item{tagCpuGpu}{integer, indicator for CPU or GPU computations (see
Details).}
\item{tag1D2D}{integer, indicator regarding data partitioning for
parallelization (see Details).}
\item{device_id}{integer, id of GPU device on the machine (see Details).}
}
\description{
When using compiled user-defined operators, \code{rkeops} requires runtime
options that control how the computations are done (memory manegement,
data partition for parallelization, etc.).

The function \code{runtime_options} returns an object of class
\code{rkeops_runtime_options} with user-provided values for the corresponding
options (see Details).
}
\details{
The aforementioned compile options are the following:
\itemize{
\item \code{tagCpuGpu}: \code{0} means convolution on Cpu, \code{1} means convolution on Gpu,
\code{2} means convolution on Gpu using data on device (i.e. in GPU memory).
Default value is \code{1}.
\item \code{tag1D2D}: \code{0} means 1D parallelization (over rows of matrices), and \code{1}
parallelization over blocks of rows and columns (useful with small columns
large rows matrices). Default value is \code{0}.
\item \code{device_id}: id of GPU device on the machine where the computation are
made. Default value is \code{0}. Ideally, GPU assignation should be handled
outside of \code{R} and \code{rkeops}.
}

\strong{Note:} To set up the options defined with this function, you should use
the functions \code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}} or
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}.
}
\seealso{
\code{\link[rkeops:default_runtime_options]{rkeops::default_runtime_options()}}, \code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}},
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}
}
\author{
Ghislain Durif
}
