% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{ScalarProduct.or.OR}
\alias{ScalarProduct.or.OR}
\alias{|}
\title{Euclidean scalar product (for LazyTensors) or default logical "or"}
\usage{
x | y
}
\arguments{
\item{x, y}{input for \code{\link[=|.default]{|.default()}} or \code{\link[=|.LazyTensor]{|.LazyTensor()}}.}
}
\value{
See value of \code{\link[=|.default]{|.default()}} or \code{\link[=|.LazyTensor]{|.LazyTensor()}}.
}
\description{
Euclidean scalar product symbolic binary operation for \code{LazyTensor} objects,
i.e. \eqn{\langle x, y\rangle}, or default element-wise logical "or" for
other types, i.e. \verb{x OR y}.
}
\details{
If \code{x} or \code{y} is a \code{LazyTensor}, see \code{\link[=|.LazyTensor]{|.LazyTensor()}}, else see
\code{\link[=|.default]{|.default()}}.
}
\examples{
# R base element-wise logical or operation
TRUE | FALSE
x <- 1:10
(x < 2) | (x > 8)
# LazyTensor symbolic scalar product operation
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
y <- matrix(runif(250 * 3), 250, 3) # arbitrary R matrix, 250 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
y_j <- LazyTensor(y, index = 'j')   # creating LazyTensor from matrix y, 
                                    # indexed by 'j'

x_sp_y <- x_i | y_j                 # symbolic matrix
}
\seealso{
\code{\link[=|.default]{|.default()}}, \code{\link[=|.LazyTensor]{|.LazyTensor()}}
}
