% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keops_setup.R
\name{set_pykeops_verbosity}
\alias{set_pykeops_verbosity}
\title{Set PyKeOps verbosity}
\usage{
set_pykeops_verbosity()
}
\value{
None
}
\description{
Enable or disable verbosity during operator compilation process in PyKeops
using current verbosity status in RKeOps options.
}
\details{
\strong{Note:} this is an internal function, user should use
\code{\link[=rkeops_enable_verbosity]{rkeops_enable_verbosity()}} or \code{\link[=rkeops_disable_verbosity]{rkeops_disable_verbosity()}}.
}
\examples{
\dontrun{
set_pykeops_verbosity()
}
}
\seealso{
\code{\link[=setup_pykeops]{setup_pykeops()}}
}
\author{
Ghislain Durif
}
\keyword{internal}
