% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_options.R
\name{set_rkeops_cache_dir}
\alias{set_rkeops_cache_dir}
\title{Set RKeOps cache directory}
\usage{
set_rkeops_cache_dir(cache_dir = NULL, verbose = TRUE)
}
\arguments{
\item{cache_dir}{string, path to cache directory where rkeops operator
compilation byproducts will be stored to be re-used for further use (and
avoid unnecessary recompilation). Default is \code{NULL} and default build
directory (see \code{\link[=default_rkeops_cache_dir]{default_rkeops_cache_dir()}} is used. Otherwise
\code{cache_dir} should be an existing writable directory on the system.}

\item{verbose}{logical, enables verbosity or not. Default is \code{TRUE}.}
}
\value{
None
}
\description{
Set the path to the cache folder where rkeops operator
compilation byproducts will be stored to be re-used for further use
(and avoid unnecessary recompilation).
}
\details{
When compiling a user-defined operators or running a LazyTensor-based
operations, a shared object (\code{.so}) library (or dynamic link library,
\code{.dll}) file is created in RKeOps cache directory.

For every new operators or new LazyTensor operations, such a file is
created, to be reused next time the same computations is done (then saving
compilation time).

Calling \code{get_rkeops_cache_dir()} gives the location of this specific
directory on your system.

You can use \code{\link[=stat_rkeops_cache_dir]{stat_rkeops_cache_dir()}} to verify rkeops cache
directory disk usage, and you can use \code{\link[=clean_rkeops]{clean_rkeops()}}
to (fully or partially) delete its content.

\strong{Note:} see \code{\link[=default_rkeops_cache_dir]{default_rkeops_cache_dir()}} for more details about
the default rkeops cache directory.
}
\examples{
# use default cache directory
set_rkeops_cache_dir()
\dontrun{
# edit path at your convenience
set_rkeops_cache_dir("/path/to/a/specific/directory")
}

}
\seealso{
\code{\link[=get_rkeops_cache_dir]{get_rkeops_cache_dir()}}, \code{\link[=stat_rkeops_cache_dir]{stat_rkeops_cache_dir()}},
\code{\link[=clean_rkeops]{clean_rkeops()}}
}
\author{
Ghislain Durif
}
