% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-set_options.R
\name{set_rkeops_option}
\alias{set_rkeops_option}
\title{Set up a specific compile or runtime options of \code{rkeops} in \code{R} global
options scope}
\usage{
set_rkeops_option(option, value)
}
\arguments{
\item{option}{string, name of the options to set up (see Details).}

\item{value}{whatever value to assign to the chosen option (see Details).}
}
\description{
The function \code{set_rkeops_option} allows to set up a single specific \code{rkeops}
options in \code{R} global options scope.
}
\details{
\code{rkeops} uses two sets of options: compile options \code{build_dir}, \code{src_dir},
\code{precision}, \code{verbosity}, \code{use_cuda_if_possible} (see
\code{\link[rkeops:compile_options]{rkeops::compile_options()}}), and runtime options \code{tagCpuGpu}, \code{tag1D2D},
\code{device_id} (see \code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}). These options define the
behavior of \code{rkeops} when compiling or when running new user-defined
operators.

With the function \code{set_rkeops_option}, you can set up a specific \code{rkeops}
options among \code{build_dir}, \code{src_dir}, \code{precision}, \code{verbosity},
\code{use_cuda_if_possible} or \code{tagCpuGpu}, \code{tag1D2D}, \code{device_id} with a value
that you provide in input.

To know which values are allowed for which options, you can check
\code{\link[rkeops:compile_options]{rkeops::compile_options()}} and \code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}.
}
\seealso{
\code{\link[rkeops:set_rkeops_options]{rkeops::set_rkeops_options()}}, \code{\link[rkeops:compile_options]{rkeops::compile_options()}},
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}
}
\author{
Ghislain Durif
}
