% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keops_setup.R
\name{setup_pykeops}
\alias{setup_pykeops}
\title{Setup PyKeOps before using it}
\usage{
setup_pykeops()
}
\value{
None
}
\description{
Wrapper function to call all setup functions for PyKeOps including:
\itemize{
\item \code{\link[=set_pykeops_verbosity]{set_pykeops_verbosity()}} to set verbosity
\item \code{\link[=set_pykeops_build_dir]{set_pykeops_build_dir()}} to set build directory
}
}
\details{
\strong{Note:} this is an internal function, user should use
\code{\link[=rkeops_enable_verbosity]{rkeops_enable_verbosity()}} or \code{\link[=rkeops_disable_verbosity]{rkeops_disable_verbosity()}}
for verbosity setup, \code{\link[=get_rkeops_cache_dir]{get_rkeops_cache_dir()}} or
\code{\link[=set_rkeops_cache_dir]{set_rkeops_cache_dir()}} for cache/build directory setup.
}
\examples{
\dontrun{
setup_pykeops()
}
}
\seealso{
\code{\link[=set_pykeops_verbosity]{set_pykeops_verbosity()}}, \code{\link[=set_pykeops_build_dir]{set_pykeops_build_dir()}}
}
\author{
Ghislain Durif
}
\keyword{internal}
