% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_req.R
\name{setup_pyreq}
\alias{setup_pyreq}
\title{Setup Python requirements for RKeOps}
\usage{
setup_pyreq()
}
\description{
Install requirements (i.e. Python version and Python packages) for
RKeOps to work.
}
\details{
RKeOps now uses PyKeOps Python package under the hood thanks to the
\href{https://rstudio.github.io/reticulate/}{\code{reticulate}} R package that
provides an "R Interface to Python".

This function is a wrapper around the \code{\link[reticulate:py_require]{reticulate::py_require()}}
function that allows to setup a temporary Python environment on the fly
for the current R session.

\strong{Important 1:} for the moment, Python version requirement is hard-coded
to be 3.12. This could change in the future depending on PyKeOps evolution.

\strong{Important 1:} Python is a requirement as an intern machinery for the
package to work but you will not need to create nor manipulate Python
codes to use the RKeOps package.
}
\author{
Ghislain Durif
}
