% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sign.default}
\alias{sign.default}
\title{Sign function}
\usage{
\method{sign}{default}(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
See 'Details' section.
}
\description{
\code{sign} returns a vector with the signs of the corresponding
  elements of \code{x} (the sign of a real number is 1, 0, or \eqn{-1}
  if the number is positive, zero, or negative, respectively).

  Note that \code{sign} does not operate on complex vectors.
}
\details{
This is an \link[base]{internal generic} \link[base]{primitive} function: methods
  can be defined for it directly or via the
  \code{\link[base:S3groupGeneric]{Math}} group generic.
}
\examples{
sign(pi)    # == 1
sign(-2:3)  # -1 -1 0 1 1 1
}
\seealso{
\code{\link[base:sign]{base::sign()}}
}
\author{
R core team and contributors
}
