% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sign.LazyTensor}
\alias{sign.LazyTensor}
\title{Element-wise sign operation}
\usage{
\method{sign}{LazyTensor}(x)
}
\arguments{
\item{x}{a \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values,
or a scalar value.}
}
\value{
An object of class \code{LazyTensor}.
}
\description{
Symbolic element-wise sign operation for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{sign(x)} returns a \code{LazyTensor} that
encodes, symbolically, the element-wise sign of \code{x}, i.e.
\deqn{\text{sign}(x) = \left\{
\begin{array}{rcl}
0 & \text{if } & x = 0\\
-1 & \text{if } & x < 0\\
1 & \text{if } & x > 0 
\end{array}\right.}
}
\examples{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Sign_x <- sign(x_i)                 # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=sign]{sign()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
