% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sin}
\alias{sin}
\title{Element-wise sine operation.}
\usage{
sin(x)
}
\arguments{
\item{x}{input for \code{\link[=sin.default]{sin.default()}} or
\code{\link[=sin.LazyTensor]{sin.LazyTensor()}}.}
}
\value{
See value of \code{\link[=sin.default]{sin.default()}} or
\code{\link[=sin.LazyTensor]{sin.LazyTensor()}}.
}
\description{
Symbolic element-wise sine operation for \code{LazyTensor}
objects or standard element-wise sine operation otherwise.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=sin.LazyTensor]{sin.LazyTensor()}}, else
see \code{\link[=sin.default]{sin.default()}}.
}
\examples{
# R base operation
sin(0)
sin(pi)
# LazyTensor symbolic element-wise sine
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Sin_x <- sin(x_i)                   # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=sin.default]{sin.default()}}, \code{\link[=sin.LazyTensor]{sin.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
