% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sinxdivx}
\alias{sinxdivx}
\title{Element-wise \code{sin(x)/x} function}
\usage{
sinxdivx(x)
}
\arguments{
\item{x}{input for \code{\link[=sinxdivx.default]{sinxdivx.default()}} or
\code{\link[=sinxdivx.LazyTensor]{sinxdivx.LazyTensor()}}.}
}
\value{
See value of \code{\link[=sinxdivx.default]{sinxdivx.default()}} or
\code{\link[=sinxdivx.LazyTensor]{sinxdivx.LazyTensor()}}
}
\description{
Symbolic element-wise \code{sin(x)/x} function for \code{LazyTensor} objects or
standard element-wise \code{sin(x)/x} function otherwise.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=sinxdivx.LazyTensor]{sinxdivx.LazyTensor()}},
else see \code{\link[=sinxdivx.default]{sinxdivx.default()}}.

\strong{Note}: by convention \code{sinxdivx(0)} returns \code{1}.
}
\examples{
# Numerical input
sinxdivx(4)
sinxdivx(1:10)
# LazyTensor symbolic element-wise `sin(x)/x`
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
sindiv_x <- sinxdivx(x_i)           # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=sinxdivx.default]{sinxdivx.default()}}, \code{\link[=sinxdivx.LazyTensor]{sinxdivx.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
