% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{/.ComplexLazyTensor}
\alias{/.ComplexLazyTensor}
\title{Division.}
\usage{
\method{/}{ComplexLazyTensor}(x, y)
}
\description{
Division.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
