% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stat_dir}
\alias{stat_dir}
\title{Disk usage of a given directory}
\usage{
stat_dir(path)
}
\arguments{
\item{path}{character string, path to directory for which the disk
usage will be computed.}
}
\value{
character string, a disk usage in (G/M/K)bytes.
}
\description{
Sum the sizes of all files in a given folder and its sub-directories.
}
\author{
Ghislain Durif
}
\keyword{internal}
