% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-string2hash.R
\name{string2hash}
\alias{string2hash}
\title{Generatation of hash values from a text strings}
\usage{
string2hash(str)
}
\arguments{
\item{str}{a text string.}
}
\value{
the associated hash value as a character
string.
}
\description{
This function uses a sha256 algorithm to generate
a hash from a text string.
}
\details{
This function is used to generate unique identifier
from text formulae that is used to store
compiled associated code.
}
\author{
Ghislain Durif
}
\keyword{internal}
