% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sum.default}
\alias{sum.default}
\title{Summation operation.}
\usage{
\method{sum}{default}()
}
\description{
Summation operation.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
