% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sumsoftmaxweight}
\alias{sumsoftmaxweight}
\alias{sumsoftmaxweight_reduction}
\title{Sum of weighted Soft-Max reduction.}
\usage{
sumsoftmaxweight(x, index, weight)

sumsoftmaxweight_reduction(x, index, weight)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{a \code{character} corresponding to the reduction dimension that
should be either \code{"i"} or \code{"j"} to specify whether if the reduction is
indexed by \code{"i"} or \code{"j"}.}

\item{weight}{an optional object (\code{LazyTensor} or \code{ComplexLazyTensor}) that
specifies scalar or vector-valued weights.}
}
\value{
a matrix corresponding to the Sum of weighted Soft-Max reduction.
}
\description{
Sum of weighted Soft-Max reduction.

Sum of weighted Soft-Max reduction.
}
\details{
If \code{x} is a \code{LazyTensor} or a \code{ComplexLazyTensor},
\code{sumsoftmaxweight(x, index, weight)} will:
\itemize{
\item if \code{index = "i"}, return the Sum of weighted Soft-Max reduction of \code{x}
over the \code{i} indexes;
\item if \code{index = "j"}, return the Sum of weighted Soft-Max reduction of \code{x}
over the \code{j} indexes.
}

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Reductions"}.
}
\examples{
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) 
x_i <- LazyTensor(x, index = 'i') 
y <- matrix(runif(100 * 3), 100, 3)
y_j <- LazyTensor(y, index = 'j')

V_ij <- x_i - y_j   # weight matrix
S_ij = sum(V_ij^2)     

ssmaxweight <- sumsoftmaxweight(S_ij, 'i', V_ij) # sumsoftmaxweight reduction
                                                 # over the 'i' indices
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
\donttest{
\dontshow{
# Specific setup for example runs (not necessary for standard use)
reticulate::py_available(initialize = TRUE)
set_rkeops_options()
# For build on CRAN, use CPU computing with 2 cores max
rkeops_use_cpu(ncore = 2)
}
x <- matrix(runif(150 * 3), 150, 3) 
x_i <- LazyTensor(x, index = 'i') 
y <- matrix(runif(100 * 3), 100, 3)
y_j <- LazyTensor(y, index = 'j')

V_ij <- x_i - y_j   # weight matrix
S_ij = sum(V_ij^2)     

# sumsoftmaxweight reduction over the 'i' indices
ssmaxw_red <- sumsoftmaxweight_reduction(S_ij, 'i', V_ij) 
\dontshow{
# clean-up (for CRAN build)
clean_rkeops(remove_cache_dir = TRUE)
}
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
