% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_gpu}
\alias{use_gpu}
\title{Enable GPU-computing when calling user-defined operators}
\usage{
use_gpu()
}
\description{
Set up \code{rkeops} runtime options to use GPU computing when calling
user-defined operators.
}
\details{
If you have compiled GPU-compatible operators (see \code{\link[rkeops:compile4gpu]{rkeops::compile4gpu()}}),
you can call the function \code{use_gpu} to specificly run computations on GPU.

\strong{Note:} The default behavior in \code{rkeops} is to use CPU computing, thus
calling the function \code{use_gpu} is mandatory to run computations on GPU.
}
\examples{
\dontrun{
use_gpu()
}
}
\seealso{
\code{\link[rkeops:compile4gpu]{rkeops::compile4gpu()}}
}
\author{
Ghislain Durif
}
