% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{xlogx.default}
\alias{xlogx.default}
\title{x*log(x) function.}
\usage{
\method{xlogx}{default}(x)
}
\description{
x*log(x) function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
