% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{xlogx}
\alias{xlogx}
\title{Element-wise \code{x*log(x)} operation}
\usage{
xlogx(x)
}
\arguments{
\item{x}{input for \code{\link[=xlogx.default]{xlogx.default()}} or
\code{\link[=xlogx.LazyTensor]{xlogx.LazyTensor()}}.}
}
\value{
See value of \code{\link[=xlogx.default]{xlogx.default()}} or
\code{\link[=xlogx.LazyTensor]{xlogx.LazyTensor()}}
}
\description{
Symbolic element-wise \code{x*log(x)} function for \code{LazyTensor} objects or
standard element-wise \code{x*log(x)} function otherwise.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=xlogx.LazyTensor]{xlogx.LazyTensor()}},
else see \code{\link[=xlogx.default]{xlogx.default()}}.

\strong{Note}: by convention \code{xlogx(0)} returns \code{0}.
}
\examples{
# Numerical input
xlogx(4)
xlogx(1:10)
# LazyTensor symbolic element-wise `x*log(x)`
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
xlog_x <- xlogx(x_i)                # symbolic matrix, 150 rows and 3 columns
}
\seealso{
\code{\link[=xlogx.default]{xlogx.default()}}, \code{\link[=xlogx.LazyTensor]{xlogx.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
