% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{abs}
\alias{abs}
\title{Element-wise absolute value (or modulus) operation}
\usage{
abs(x)
}
\arguments{
\item{x}{input for \code{\link[=abs.default]{abs.default()}} or \code{\link[=abs.LazyTensor]{abs.LazyTensor()}}.}
}
\value{
See value of \code{\link[=abs.default]{abs.default()}} or \code{\link[=abs.LazyTensor]{abs.LazyTensor()}}.
}
\description{
Symbolic element-wise absolute value (or modulus) operation for
\code{LazyTensor} objects or default element-wise absolute value (or modulus)
operation otherwise, i.e. \eqn{\vert x \vert}.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=abs.LazyTensor]{abs.LazyTensor()}}, else
see \code{\link[=abs.default]{abs.default()}}.

\strong{Note}: for complex input argument \code{z}, then the absolute value
corresponds to the complex modulus \code{abs(z) = Mod(z)}.
}
\examples{
\dontrun{
# R base operation
abs(5)
abs(-5)
abs(-10:10)
# LazyTensor symbolic element-wise absolute value
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Abs_x <- abs(x_i)                   # symbolic matrix, 150 rows and 3 columns
}
}
\seealso{
\code{\link[=abs.default]{abs.default()}}, \code{\link[=abs.LazyTensor]{abs.LazyTensor()}},
\code{\link[=Mod.default]{Mod.default()}}, \code{\link[=Mod.LazyTensor]{Mod.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
