% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{acos.LazyTensor}
\alias{acos.LazyTensor}
\title{Element-wise arc-cosine.}
\usage{
\method{acos}{LazyTensor}(x)
}
\description{
Element-wise arc-cosine.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
