% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{argmax}
\alias{argmax}
\alias{argmax_reduction}
\title{ArgMax operation or ArgMax reduction}
\usage{
argmax(x, index = NA)

argmax_reduction(x, index)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{a \code{character} corresponding to the wanted reduction dimension,
either \code{"i"} or \code{"j"}, to specify whether if the argmax reduction is done
along the index \code{i} or \code{j}.
It can be \code{NA} (default) when no reduction is desired.}
}
\value{
a \code{LazyTensor} if \code{index = NA} or an array storing the result of
the specified argmax reduction otherwise.
}
\description{
ArgMax operation or ArgMax reduction for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{argmax(x, index)} will :
\itemize{
\item if \code{index = "i"}, return the argmax reduction of \code{x} over the \code{i}
indexes.
\item if \code{index = "j"}, return the argmax reduction of \code{x} over the \code{j}
indexes.
\item if \code{index = NA} (default), return a new \code{LazyTensor} object
representing the symbolic argmax of the values along the inner dimension
of \code{x}.
}

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Reductions"}.

\strong{Note}: \code{index} input argument cannot be \code{NA} for the
\code{argmax_reduction()} function.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

argmax_xi <- argmax(x_i, "i")  # argmax reduction indexed by 'i'
argmax_x <- argmax(x_i)        # symbolic matrix
}
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

argmax_red <- argmax_reduction(x_i, "i")  # argmax reduction indexed by 'i'
}
}
\seealso{
\code{\link[=argmax_reduction]{argmax_reduction()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
