% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{argmax_reduction}
\alias{argmax_reduction}
\title{ArgMax reduction.}
\usage{
argmax_reduction(x, index)
}
\arguments{
\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{A \code{character} corresponding to the reduction dimension that
should be either \strong{i} or \strong{j} to specify whether if the reduction is
indexed by \strong{i} (rows) or \strong{j} (columns).}
}
\value{
A matrix corresponding to the argmax reduction.
}
\description{
ArgMax reduction.
}
\details{
\code{argmax_reduction(x, index)} will return the argmax reduction of
\code{x}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

argmax_red <- argmax_reduction(x_i, "i")  # argmax reduction indexed by "i"
}
}
\seealso{
\code{\link[=argmax]{argmax()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
