% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{argmin}
\alias{argmin}
\title{ArgMin.}
\usage{
argmin(x, index = NA)
}
\arguments{
\item{x}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector or a matrix of numeric values, or a scalar value.}

\item{index}{A \code{character} corresponding to the reduction dimension that should
be either \strong{i} or \strong{j} to specify whether if the summation is indexed by \strong{i} (rows),
or \strong{j} (columns).
It can be NA (default) when no reduction is desired.}
}
\value{

}
\description{
ArgMin unary operation, or ArgMin reduction.
}
\details{
\code{argmin(x, index)} will :
\itemize{
\item if \strong{index = "i"}, return the argmin reduction of \strong{x} over the "i" indexes.
\item if \strong{index = "j"}, return the argmin reduction of \strong{x} over the "j" indexes.
\item if \strong{index = NA} (default), return a new \code{LazyTensor} object representing
the argmin  of the values of the vector.
}
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'

argmin_xi <- argmin(x_i, "i")  # argmin reduction indexed by "i"
argmin_x <- argmin(x_i)        # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
