% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{arithmetic.add}
\alias{arithmetic.add}
\alias{+}
\title{Addition}
\usage{
\special{+x}
\special{x + y}
}
\arguments{
\item{x, y}{input for \code{\link[=+.default]{+.default()}} or \code{\link[=+.LazyTensor]{+.LazyTensor()}}.}
}
\value{
See value of \code{\link[=+.default]{+.default()}} or \code{\link[=+.LazyTensor]{+.LazyTensor()}}.
}
\description{
Symbolic addition for \code{LazyTensor} objects or default addition for other
types.
}
\details{
If \code{x} or \code{y} is a \code{LazyTensor}, see \code{\link[=+.LazyTensor]{+.LazyTensor()}}, else see
\code{\link[=+.default]{+.default()}}.

\strong{Note:}
\itemize{
\item the \code{+} operand is only a binary operator for \code{LazyTensor}s: \code{x + y}.
\item the \code{+} operand can be a unary or a binary operator for other types: \code{+ x}
or \code{x + y}.
}
}
\examples{
\dontrun{
# R base operation
+5
1 + 3
# LazyTensor symbolic addition
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
y <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x,
                                    # indexed by 'i'
y_j <- LazyTensor(y, index = 'j')   # creating LazyTensor from matrix y,
                                    # indexed by 'j'
Sum_xy <- x_i + y_j                 # symbolic matrix, 150 rows and 3 columns
}
}
\seealso{
\code{\link[=+.default]{+.default()}}, \code{\link[=+.LazyTensor]{+.LazyTensor()}},
\code{\link[=+.ComplexLazyTensor]{+.ComplexLazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
