% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{arithmetic.divide}
\alias{arithmetic.divide}
\alias{/}
\title{Division}
\usage{
x / y
}
\arguments{
\item{x, y}{input for \code{\link[=/.default]{/.default()}} or \code{\link[=/.LazyTensor]{/.LazyTensor()}}.}
}
\value{
See value of \code{\link[=/.default]{/.default()}} or \code{\link[=/.LazyTensor]{/.LazyTensor()}}.
}
\description{
Symbolic division for \code{LazyTensor} objects or default division
otherwise.
}
\details{
If \code{x} or \code{y} is a \code{LazyTensor}, see \code{\link[=/.LazyTensor]{/.LazyTensor()}}, else see
\code{\link[=/.default]{/.default()}}.
}
\examples{
\dontrun{
# R base operation
+5
1 + 3
# LazyTensor symbolic division
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
y <- matrix(runif(250 * 3), 250, 3) # arbitrary R matrix, 250 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
y_j <- LazyTensor(y, index = 'j')   # creating LazyTensor from matrix y,
                                    # indexed by 'j'
x_div_y <- x_i / y_j                # symbolic matrix
}
}
\seealso{
\code{\link[=/.default]{/.default()}}, \code{\link[=/.LazyTensor]{/.LazyTensor()}},
\code{\link[=/.ComplexLazyTensor]{/.ComplexLazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
