% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{asin.LazyTensor}
\alias{asin.LazyTensor}
\title{Element-wise arc-sine.}
\usage{
\method{asin}{LazyTensor}(x)
}
\description{
Element-wise arc-sine.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
