% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{atan.LazyTensor}
\alias{atan.LazyTensor}
\title{Element-wise arc-tangent.}
\usage{
\method{atan}{LazyTensor}(x)
}
\description{
Element-wise arc-tangent.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
