% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{atan}
\alias{atan}
\title{Element-wise arctangent.}
\usage{
atan(x)
}
\arguments{
\item{x}{A \code{LazyTensor}, a vector or a matrix of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor" if the function is called with a \code{LazyTensor},
and an object of class "numeric", "matrix", or "array" otherwise, depending on the input class
(see R default \code{atan()} function).
}
\description{
Symbolic unary operation for element-wise arctangent.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{atan(x)} returns a \code{LazyTensor} that encodes, symbolically,
the element-wise arctangent of \code{x} ; else, computes R default arctangent function.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'
Atan_x <- atan(x_i)                 # symbolic matrix, 150 rows and 3 columns
}
}
\author{
Ghislain Durif
}
