% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{atan}
\alias{atan}
\title{Element-wise arc-tangent operation.}
\usage{
atan(x)
}
\arguments{
\item{x}{input for \code{\link[=atan.default]{atan.default()}} or
\code{\link[=atan.LazyTensor]{atan.LazyTensor()}}.}
}
\value{
See value of \code{\link[=atan.default]{atan.default()}} or
\code{\link[=atan.LazyTensor]{atan.LazyTensor()}}.
}
\description{
Symbolic element-wise arc-tangent operation for \code{LazyTensor}
objects or standard element-wise arc-tangent operation otherwise.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=atan.LazyTensor]{atan.LazyTensor()}}, else
see \code{\link[=atan.default]{atan.default()}}.
}
\examples{
\dontrun{
# R base operation
atan(0)
atan(-1)
# LazyTensor symbolic element-wise arc-tangent
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Atan_x <- atan(x_i)                 # symbolic matrix, 150 rows and 3 columns
}
}
\seealso{
\code{\link[=atan.default]{atan.default()}}, \code{\link[=atan.LazyTensor]{atan.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
