% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{atan2.LazyTensor}
\alias{atan2.LazyTensor}
\title{Element-wise 2-argument arc-tangent.}
\usage{
\method{atan2}{LazyTensor}(x, y)
}
\description{
Element-wise 2-argument arc-tangent.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
