% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{binaryop.LazyTensor}
\alias{binaryop.LazyTensor}
\title{Build a binary operation}
\usage{
binaryop.LazyTensor(x, y, opstr, is_operator = FALSE)
}
\arguments{
\item{x}{A LazyTensor, a vector of numeric values, or a scalar value.}

\item{y}{A LazyTensor, a vector of numeric values, or a scalar value.}

\item{opstr}{A text string corresponding to an operation.}

\item{is_operator}{A boolean used to specify if \strong{opstr} is an operator like \code{+}
, \code{-} or a "genuine" function.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolically applies \strong{opstr} operation to \strong{x} and \strong{y}.
}
\examples{
\dontrun{
}
}
\author{
Ghislain Durif
}
