% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{check_index}
\alias{check_index}
\title{Check index.}
\usage{
check_index(index)
}
\arguments{
\item{index}{to check.}
}
\value{
A boolean TRUE or FALSE.
}
\description{
Checks index for operation.
}
\details{
\code{check_index(index)} will return a boolean to check if \code{index} is
a character and corresponding to \code{"i"} or \code{"j"}.
\itemize{
\item if \code{index = "i"}, return \code{TRUE}.
\item if \code{index = "j"}, return \code{TRUE}.
\item else return \code{FALSE}.
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
