% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_check.R
\name{check_rkeops}
\alias{check_rkeops}
\title{Check if \code{rkeops} is ready and working}
\usage{
check_rkeops(warn = TRUE)
}
\arguments{
\item{warn}{boolean, if TRUE (default), warn user about check result.}
}
\value{
boolean value indicating if the \code{rkeops} package is ready.
}
\description{
WRITE ME
}
\examples{
\dontrun{
check_rkeops()
}
}
