% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_check.R
\name{check_rkeops}
\alias{check_rkeops}
\title{Check if \code{rkeops} is ready and working}
\usage{
check_rkeops(warn = TRUE)
}
\arguments{
\item{warn}{boolean, if TRUE (default), warn user about check result.}
}
\value{
boolean value indicating if the \code{rkeops} package is ready.
}
\description{
The function checks if the \code{rkeops} package is working and ready for use.
}
\details{
Under the hood, several verifications are made:
\itemize{
\item assess if Python is available on the system
\item assess if \code{rkeops} Python package dependencies are available (namely
\code{keopscore} and \code{pykeops}, that should be installed along with \code{rkeops})
\item assess if \code{rkeops} internal machinery is working
}
}
\examples{
\dontrun{
check_rkeops()
}
}
\author{
Ghislain Durif
}
