% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{clamp}
\alias{clamp}
\title{Element-wise clamp function.}
\usage{
clamp(x, y, z)
}
\arguments{
\item{x}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}

\item{y}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}

\item{z}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolic ternary operation for element-wise clamp function.
}
\details{
\code{clamp(x, y, z)} returns a \code{LazyTensor} that encodes, symbolically,
the element-wise clamping of \code{x} in \verb{(y, z)}.
Broadcasting rules apply.
}
\examples{
\dontrun{
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
