% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_rkeops}
\alias{clean_rkeops}
\title{Clean build directory}
\usage{
clean_rkeops()
}
\description{
Remove all shared object files generated from compilations of user-defined
operators.
}
\details{
FIXME
}
\author{
Ghislain Durif
}
