% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_rkeops}
\alias{clean_rkeops}
\title{Clean build directory}
\usage{
clean_rkeops()
}
\value{
None
}
\description{
Remove all dll files generated from compilations of user-defined
operators.
}
\details{
When compiling a user-defined operators, a shared object (so) library
(or dynamic link library, dll) file is created in the directory \code{build_dir}
specified in compile options of \code{rkeops}. For every new operators, such a
file is created.

Calling \code{clean_rkeops()} allows you to empty the directory \code{build_dir}.
}
\examples{
library(rkeops)
clean_rkeops()
}
\seealso{
\code{\link[=compile_options]{compile_options()}}, \code{\link[=set_rkeops_option]{set_rkeops_option()}}
}
\author{
Ghislain Durif
}
