% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{cos.LazyTensor}
\alias{cos.LazyTensor}
\title{Element-wise cosine.}
\usage{
\method{cos}{LazyTensor}(x)
}
\description{
Element-wise cosine.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
