% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{cos}
\alias{cos}
\title{Element-wise cosine.}
\usage{
cos(x)
}
\arguments{
\item{x}{A LazyTensor, a vector of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor" if the function is called with a \code{LazyTensor},
and an object of class "numeric", "matrix", or "array" otherwise, depending on the input class
(see R default \code{cos()} function).
}
\description{
Symbolic unary operation for element-wise cosine.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{exp(x)} returns a \code{LazyTensor} that encodes, symbolically,
the element-wise cosine of \code{x} ; else, computes R default cosine.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'
Cos_x <- cos(x_i)                   # symbolic matrix, 150 rows and 3 columns
}
}
\author{
Ghislain Durif
}
