% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{default.math.fun}
\alias{default.math.fun}
\alias{sqrt.default}
\alias{abs.default}
\title{Miscellaneous Mathematical Functions}
\usage{
\method{sqrt}{default}(x)

\method{abs}{default}(x)
}
\arguments{
\item{x}{a numeric or \code{\link[base]{complex}} vector or array.}
}
\description{
\code{abs(x)} computes the absolute value of x, \code{sqrt(x)} computes the
  (principal) square root of x, \eqn{\sqrt{x}}.% Details for complex x are below

  The naming follows the standard for computer languages such as C or Fortran.
}
\details{
These are \link[base]{internal generic} \link[base]{primitive} functions: methods
  can be defined for them individually or via the
  \code{\link[base:S3groupGeneric]{Math}} group generic.  For complex
  arguments (and the default method), \code{z}, \code{abs(z) ==
  \link[base]{Mod}(z)} and \code{sqrt(z) == z^0.5}.

  \code{abs(x)} returns an \code{\link[base]{integer}} vector when \code{x} is
  \code{integer} or \code{\link[base]{logical}}.
}
\examples{
require(stats) # for spline
require(graphics)
xx <- -9:9
plot(xx, sqrt(abs(xx)),  col = "red")
lines(spline(xx, sqrt(abs(xx)), n=101), col = "pink")
}
\seealso{
\code{\link[base:MathFun]{base::sqrt()}}

\code{\link[base:MathFun]{base::abs()}}
}
\author{
R core team and contributors
}
