% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_options.R
\name{default_rkeops_cache_dir}
\alias{default_rkeops_cache_dir}
\title{Default cache directory for RKeOps}
\usage{
default_rkeops_cache_dir(create = TRUE)
}
\arguments{
\item{create}{logical, indicate if the cache directory should be created or
not. Default is \code{TRUE}.}
}
\value{
string, path to default rkeops cache directory.
}
\description{
Default value for the path to the cache folder where rkeops operator
compilation byproducts will be stored to be re-used for further use
(and avoid unnecessary recompilation).
}
\details{
When compiling a user-defined operators or running a LazyTensor-based
operations, a shared object (\code{.so}) library (or dynamic link library,
\code{.dll}) file is created in RKeOps cache directory.

For every new operators or new LazyTensor operations, such a file is
created, to be reused next time the same computations is done (then saving
compilation time).

Calling \code{default_rkeops_cache_dir()} returns the default value for the
location of this specific directory on your system, which is:
\itemize{
\item \code{tools::R_user_dir(package = "rkeops", which = "cache")} \strong{for R>4.0}
(see \code{\link[tools:userdir]{tools::R_user_dir()}} for more details).
\item a temporary directory \strong{that is erased when R is closed} (hence losing
all compilation byproducts that will not be available anymore)
\strong{for R<4.0}.
}

You can use \code{\link[=stat_rkeops_cache_dir]{stat_rkeops_cache_dir()}} to verify rkeops cache
directory disk usage, and you can use \code{\link[=clean_rkeops]{clean_rkeops()}}
to (fully or partially) delete its content.
}
\keyword{internal}
