% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{elem}
\alias{elem}
\title{Elem.}
\usage{
elem(x, m)
}
\arguments{
\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{m}{An \code{integer} corresponding to the index of the element of \code{x} we
want to extract (\code{x[m]}).}
}
\value{
A \code{LazyTensor} that encodes, symbolically, the m-th element \code{x[m]}
of the vector \code{x}.
}
\description{
Symbolic element extraction. A unary operation.
}
\details{
Extracts the m-th element of a given vector \code{x}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # LazyTensor from matrix x, indexed by 'i'
m <- 2

elem_x <- elem(x_i, m)  # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
