% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{elemT}
\alias{elemT}
\title{ElemT.}
\usage{
elemT(x, m, n)
}
\arguments{
\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{m}{An \code{integer} corresponding at position M of a vector of zeros.}

\item{n}{An \code{integer} corresponding to the length of the vector of zeros.}
}
\value{

}
\description{
ElemT.
}
\details{
\code{elemT(x, m, n)} insert ??? TODO
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

m <- 2
n <- 3

elemT_x <- elemT(x_i, m, n)  # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
