% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{exp.LazyTensor}
\alias{exp.LazyTensor}
\title{Element-wise exponential.}
\usage{
\method{exp}{LazyTensor}(x)
}
\description{
Element-wise exponential.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
