% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{extract}
\alias{extract}
\title{Extract.}
\usage{
extract(x, m, d)
}
\arguments{
\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{m}{An \code{integer} corresponding to the starting index.}

\item{d}{An \code{integer} corresponding to the output dimension.}
}
\value{
A \code{LazyTensor} that encodes, symbolically, the sub-vector \code{x[m:m+d]}
of the vector \code{x}.
}
\description{
Symbolic sub-vector extraction. A unary operation.
}
\details{
\code{extract(x_i, m, d)} extracts a sub-vector from vector \code{x}
(\code{m} is the starting index, and \code{d} is dimension of the extracted sub-vector).
}
\examples{
\dontrun{
x <- matrix(runif(150 * 5), 150, 5) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # LazyTensor from matrix x, indexed by 'i'
m <- 2
d <- 2

extract_x <- extract(x_i, m, d)      # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
