% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{extractT}
\alias{extractT}
\title{ExtractT.}
\usage{
extractT(d)
}
\arguments{
\item{d}{An \code{integer} corresponding to the output dimension.}

\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{m}{An \code{integer} corresponding to the starting index.}
}
\value{

}
\description{
ExtractT.
}
\details{
insert vector f in a larger vector of zeros (M is starting index,
D is dimension of output)
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

m <- 2
d <- 2

extractT_x <- extractT(x_i, m, d)      # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
