% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{extractT}
\alias{extractT}
\title{ExtractT.}
\usage{
extractT(x, m, d)
}
\arguments{
\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{m}{An \code{integer} corresponding to the starting index.}

\item{d}{An \code{integer} corresponding to the output dimension.}
}
\value{

}
\description{
Insert a given range of values in a symbolic vector of zeros -
a unary operation.
}
\details{
If \code{x} is a \code{LazyTensor} encoding some values, \code{extractT(x, m, d)}
encodes, symbolically, \code{x} in a vector of zeros, starting at position \code{m}.
The output is a symbolic vector of TODO TODO TODO (M is starting index,
D is dimension of output)
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

m <- 2
d <- 2

extractT_x <- extractT(x_i, m, d)      # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
