% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_src_dir}
\alias{get_src_dir}
\title{Getter for package additional source directory}
\usage{
get_src_dir(pkg = "rkeops")
}
\arguments{
\item{pkg}{name (string) of the R package, default is "rkeops".}
}
\value{
path to the corresponding directory.
}
\description{
Return the path to the directory containing \code{keops} source
files that are shipped with \code{rkeops} installation.
}
\details{
This path is generaly something like
\verb{/path/to/R_package_install/rkeops/include}.

\strong{Note:} when running tests in the development project \code{keops} without
installing \code{rkeops}, for consistency reason, the function returns
\verb{/path/to/keops/rkeops/inst/include} (because the content of
\code{rkeops/inst/include} is copied to \code{rkeops/include} at installation).
}
\author{
Ghislain Durif
}
\keyword{internal}
