% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{ifelse.LazyTensor}
\alias{ifelse.LazyTensor}
\title{Element-wise if-else function.}
\usage{
\method{ifelse}{LazyTensor}(x, a, b)
}
\description{
Element-wise if-else function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
