% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{ifelse.LazyTensor}
\alias{ifelse.LazyTensor}
\title{Element-wise if-else function.}
\usage{
ifelse.LazyTensor(x, a, b)
}
\arguments{
\item{x}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values, or a scalar value.}

\item{a}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values, or a scalar value.}

\item{b}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolic ternary operation for element-wise if-else function.
}
\details{
\code{ifelse.LazyTensor(x, a, b)} returns a \code{LazyTensor} that encodes, symbolically,
\code{a} where \code{x >= 0} and \code{b} where \code{x < 0}.  Broadcasting rules apply.
\code{a} and \code{b} may be fixed integers or floats, or other \code{LazyTensor}.

\strong{Note}

If \code{a} and \code{b} are not scalar values, these should have the same inner dimension as \code{x}.
}
\examples{
\dontrun{
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
