% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{ifelse.LazyTensor}
\alias{ifelse.LazyTensor}
\title{Element-wise if-else function.}
\usage{
ifelse.LazyTensor(x, y, z)
}
\arguments{
\item{x}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}

\item{y}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}

\item{z}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolic ternary operation for element-wise if-else function.
}
\details{
\code{ifelse(x, y, z)} returns a \code{LazyTensor} that encodes, symbolically,
\code{y} where \code{x >= 0} and \code{z} where \code{x < 0}.  Broadcasting rules apply.
\code{y} and \code{z} may be fixed integers or floats, or other LazyTensors.
}
\examples{
\dontrun{
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
