% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{ifelse}
\alias{ifelse}
\title{Element-wise if-else function}
\usage{
ifelse(test, yes, no)
}
\arguments{
\item{test, yes, no}{input for \code{\link[=ifelse.default]{ifelse.default()}} or
\code{\link[=ifelse.LazyTensor]{ifelse.LazyTensor()}}.}
}
\value{
See value of \code{\link[=ifelse.default]{ifelse.default()}} or
\code{\link[=ifelse.LazyTensor]{ifelse.LazyTensor()}}
}
\description{
Symbolic element-wise if-else function for \code{LazyTensor} objects or
standard element-wise if-else function otherwise.
}
\details{
If \code{test} is a \code{LazyTensor}, see \code{\link[=ifelse.LazyTensor]{ifelse.LazyTensor()}}, else
see \code{\link[=ifelse.default]{ifelse.default()}}.
}
\examples{
\dontrun{
# R base operation
x <- c(6:-4)
sqrt(ifelse(x >= 0, x, NA))
# LazyTensor symbolic element-wise square
D <- 3
M <- 100
N <- 150
P <- 200
x <- matrix(runif(M * D), M, D)
y <- matrix(runif(N * D), N, D)
z <- matrix(runif(P * D), P, D)
x_i <- LazyTensor(x, index = 'i')
y_j <- LazyTensor(y, index = 'j')
z_i <- LazyTensor(z, index = 'i')

if_else_xyz <- ifelse(x_i, y_j, z_i)
}
}
\seealso{
\code{\link[=ifelse.default]{ifelse.default()}}, \code{\link[=ifelse.LazyTensor]{ifelse.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
