% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Im}
\alias{Im}
\title{Element-wise complex imaginery part operation}
\usage{
Im(z)
}
\arguments{
\item{z}{input for \code{\link[=Im.default]{Im.default()}} or \code{\link[=Im.LazyTensor]{Im.LazyTensor()}}.}
}
\value{
See value of \code{\link[=Im.default]{Im.default()}} or \code{\link[=Im.LazyTensor]{Im.LazyTensor()}}.
}
\description{
Symbolic element-wise complex imaginery part operation for
\code{ComplexLazyTensor} objects or default element-wise complex imaginery part
operation otherwise.
}
\details{
If \code{z} is a \code{LazyTensor}, see \code{\link[=Im.LazyTensor]{Im.LazyTensor()}}, else
see \code{\link[=Im.default]{Im.default()}}.
}
\examples{
\dontrun{
# R base operation
Im(1)
Im(1+1i)
# LazyTensor symbolic element-wise complex imaginery part
z <- matrix(2 + 1i^ (-6:5), nrow = 4)         # complex matrix
z_i <- LazyTensor(z, "i", is_complex = TRUE)  # creating ComplexLazyTensor
Im_z <- Im(z_i)                               # symbolic matrix
}
}
\seealso{
\code{\link[=Im.default]{Im.default()}}, \code{\link[=Im.LazyTensor]{Im.LazyTensor()}},
\code{\link[=Re]{Re()}}, \code{\link[=Arg]{Arg()}}, \code{\link[=Mod]{Mod()}}, \code{\link[=Conj]{Conj()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
