% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{index_to_int}
\alias{index_to_int}
\title{Index to int.}
\usage{
index_to_int(index)
}
\arguments{
\item{index}{A \code{character} that should be either \code{i} or \code{j}.}
}
\value{
An \code{integer}.
}
\description{
Transforms \code{string} index input into integer.
}
\details{
\code{index_to_int(index)} returns an \code{integer}: \code{1} if
\code{index == "i"} and \code{0} if \code{index == "j"}.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
