% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{inv.LazyTensor}
\alias{inv.LazyTensor}
\title{Element-wise 1/x inverse.}
\usage{
\method{inv}{LazyTensor}(x)
}
\description{
Element-wise 1/x inverse.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
