% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{is.ComplexLazyScalar}
\alias{is.ComplexLazyScalar}
\title{is.ComplexLazyScalar?}
\usage{
is.ComplexLazyScalar(x)
}
\arguments{
\item{x}{An object we want to know if it is a \code{ComplexLazyScalar}.}
}
\value{
A boolean, TRUE or FALSE.
}
\description{
Checks whether if the given input is a \code{ComplexLazyTensor}encoding
a single complex value. That is, if the input is a fixed parameter
\code{ComplexLazyTensor} of dimension 1.
}
\details{
If \code{x} is a fixed parameter \code{ComplexLazyTensor} encoding a
single complex value, \code{is.ComplexLazyScalar(x)}
returns TRUE, else, returns FALSE.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
