% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{is.ComplexLazyTensor}
\alias{is.ComplexLazyTensor}
\title{is.ComplexLazyTensor?}
\usage{
is.ComplexLazyTensor(x)
}
\arguments{
\item{x}{An object we want to know if it is a \code{ComplexLazyTensor}.}
}
\value{
A boolean, TRUE or FALSE.
}
\description{
Checks whether if the given input is a \code{ComplexLazyTensor} or not.
}
\details{
If \code{x} is a \code{ComplexLazyTensor}, \code{is.ComplexLazyTensor(x)}
returns TRUE, else, returns FALSE.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
