% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{is.int}
\alias{is.int}
\title{Scalar integer test.}
\usage{
is.int(x)
}
\arguments{
\item{x}{An object we want to know if it is a \code{integer}.}
}
\value{
A boolean, TRUE or FALSE.
}
\description{
Checks whether if the given input is a scalar \code{integer} or not.
}
\details{
If \code{x} is a scalar\code{integer}, \code{is.int(x)} returns TRUE,
else, returns FALSE.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
