% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{is.int}
\alias{is.int}
\title{Scalar integer test.}
\usage{
is.int(x)
}
\arguments{
\item{x}{An object that we want to know if it is an \code{integer}.}
}
\value{
A boolean, TRUE or FALSE.
}
\description{
Checks whether the given input is a scalar \code{integer} or not.
}
\details{
If \code{x} is a scalar\code{integer}, \code{is.int(x)} returns TRUE,
else, returns FALSE.
}
\examples{
\dontrun{
# basic example
A <- 3
B <- 3.4
C <- rep(3, 10)

is.int(A)  # returns TRUE
is.int(B)  # returns FALSE
is.int(C)  # returns FALSE
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
