% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-keops_grad.R
\name{keops_grad}
\alias{keops_grad}
\title{Compute the gradient of a rkeops operator}
\usage{
keops_grad(operator, var)
}
\arguments{
\item{operator}{a function returned by \code{keops_kernel} implementing a
formula.}

\item{var}{a text string or an integer number indicating regarding to which
variable/parameter (given by name or by position starting at 0) the
gradient of the formula should be computed.}
}
\value{
FIXME
}
\description{
FIXME
}
\details{
The use of the function \code{keops_grad} is detailled in the vignettes.
Run \code{browseVignettes("rkeops")} to access the vignettes.
FIXME
}
\examples{
\dontrun{
}
}
\author{
Ghislain Durif
}
