% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-keops_kernel.R
\name{keops_kernel}
\alias{keops_kernel}
\title{Defines a new operators}
\usage{
keops_kernel(formula, args)
}
\arguments{
\item{formula}{text string, an operator formula.}

\item{args}{vector of text string, operator parameters.}
}
\value{
FIXME
}
\description{
FIXME
}
\details{
FIXME
See \url{http://www.kernel-operations.io/keops/api/math-operations.html}
}
\examples{
\dontrun{
# Define and test a function that computes for each i the sum over j
# of the square of the scalar products of `x_i` and `y_j` (both 3d vectors)
F <- keops_kernel(formula = "Sum_Reduction(Square((x,y)),0)",
                  args = c("x=Vi(3)", "y=Vj(3)"))
x <- rnorm(3,2000)
y <- rnorm(3,5000)
res <- F(x,y)

# Define and test the convolution with a Gauss kernel i.e. the sum 
# over j of `e^(lambda*||x_i - y_j||^2) * beta_j` where `x_i`, `y_j` 
# and `beta_j` are 3d vectors
F = keops_kernel(formula = "Sum_Reduction(Exp(lambda*SqNorm2(x-y))*beta,0)",
                 args = c("x=Vi(3)", "y=Vj(3)", 
                          "beta=Vj(3)", "lambda=Pm(1)"))
x <- rnorm(3,2000)
y <- rnorm(3,5000)
beta <- rnorm(3,5000)
lambda <- 0.25
res <- F(x, y, beta, lambda)
}
}
\author{
Ghislain Durif
}
