% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{LazyTensor}
\alias{LazyTensor}
\title{Build and return a LazyTensor object}
\usage{
LazyTensor(x, index = NA)
}
\arguments{
\item{x}{A matrix or a vector of numeric values, or a scalar value}

\item{index}{A text string that should be either \code{"i"} or \code{"j"}, or an \code{NA} value (the default),
to specify whether if the \code{x} variable is indexed by i, by j, or is a fixed parameter across indices.
If \code{x} is a matrix, \code{index} must be \code{"i"} or \code{"j"}.}
}
\value{
An object of class "LazyTensor", which is a list with the following elements:

\itemize{
\item{\code{formula}:}{ A string defining the mathematical operation to be computed by the KeOps routine, of the form Var(ind,dim,cat), where :
\itemize{
\item{\code{ind}:}{ gives the position in the final call to KeOps routine}
\item{\code{dim}:}{ is the dimension of the input}
\item{\code{cat}:}{ the KeOps "category": 0 if the input is an R matrix indexed by "i", 1 if the input is an R matrix indexed by "j", or 2 if the input is a parameter vector or scalar, without any attached index}
}}
\item{\code{vars}:}{ a list of R matrices which will be the inputs of the KeOps routine}
\item{\code{ni}:}{ the number of rows of the input if it is an "i" indexed variable}
\item{\code{nj}:}{ the number of rows of the input if it is a "j" indexed variable}
}
}
\description{
LazyTensors objects are wrappers around R matrices or vectors that are used to create
symbolic formulas for the KeOps reduction operations.
}
\details{
The use of the function \code{LazyTensor} is detailed in the vignettes.
Run \code{browseVignettes("rkeops")} to access the vignettes.
}
\examples{
\dontrun{
# TODO change set_rkeops_options() below ?
set_rkeops_options()

# Data
# TODO increase `nx` and `ny`
nx <- 100
ny <- 150
x <- matrix(runif(nx * 3), nrow = nx, ncol = 3) # arbitrary R matrix representing 100 data points in R^3
y <- matrix(runif(ny * 3), nrow = ny, ncol = 3) # arbitrary R matrix representing 150 data points in R^3
s <- 0.1                                        # scale parameter

# Turn our Tensors into KeOps symbolic variables:
x_i <- LazyTensor(x, "i")   # symbolic object representing an arbitrary row of x, indexed by the letter "i"
y_j <- LazyTensor(y, "j")   # symbolic object representing an arbitrary row of y, indexed by the letter "j"

# Perform large-scale computations, without memory overflows:
D_ij <- Sum((x_i - y_j)^2)    # symbolic matrix of pairwise squared distances, with 100 rows and 150 columns
K_ij <- Exp(- D_ij / s^2)     # symbolic matrix, 100 rows and 150 columns
# TODO fix `reduction.LazyTensor`
res <- Sum(K_ij, index = "i") # actual R matrix (in fact a row vector of length 150 here)
                              # containing the column sums of K_ij
                              # (i.e. the sums over the "i" index, for each "j" index)

}
}
\author{
Ghislain Durif
}
