% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{log.LazyTensor}
\alias{log.LazyTensor}
\title{Element-wise natural logarithm.}
\usage{
\method{log}{LazyTensor}(x)
}
\description{
Element-wise natural logarithm.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
