% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{matmult.LazyTensor}
\alias{matmult.LazyTensor}
\alias{\%*\%.LazyTensor}
\title{Matrix multiplication}
\usage{
\method{\%*\%}{LazyTensor}(x, y)
}
\arguments{
\item{x, y}{a \code{LazyTensor}, a \code{ComplexLazyTensor}.}
}
\value{
A matrix.
}
\description{
Matrix multiplication binary operation for \code{LazyTensor} objects
corresponding to combination of multiplication operator \code{*} and sum
reduction.
}
\details{
If \code{x} or \code{y} is a \code{LazyTensor}, \code{x \%*\% y} returns the sum
reduction of the product \code{x * y}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
y <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
y_j <- LazyTensor(y, index = 'j')   # creating LazyTensor from matrix y, 
                                    # indexed by 'j'
x_mult_y <- x_i \%*\% y_j             # FIXME
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
