% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{matvecmult}
\alias{matvecmult}
\title{Matrix-vector product.}
\usage{
matvecmult(x, y)
}
\arguments{
\item{x}{A \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{D}{An \code{integer} corresponding to the output dimension.}
}
\value{

}
\description{
Matrix-vector product - a binary operation.
}
\details{
\code{matvecmult(x, y)} encodes, symbolically,
the matrix-vector product of \code{x} and \code{y} along their last dimension.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # LazyTensor from matrix x, indexed by 'i'

mv_mult_xy <- matvecmult(x_i, y_j) # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
