% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{max.LazyTensor}
\alias{max.LazyTensor}
\title{Maximum.}
\usage{
\method{max}{LazyTensor}(x, index = NA)
}
\description{
Maximum.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
