% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{max}
\alias{max}
\title{Maximum operation or maximum reduction}
\usage{
max(x, ...)
}
\arguments{
\item{x}{input for \code{\link[=max.default]{max.default()}} or \code{\link[=max.LazyTensor]{max.LazyTensor()}}.}

\item{...}{optional arguments for \code{\link[=max.default]{max.default()}} or
\code{\link[=max.LazyTensor]{max.LazyTensor()}}.}
}
\value{
Maximum of input.
}
\description{
Maximum operation for vector/matrix/array or \code{LazyTensor}, or maximum
reduction for \code{LazyTensor}.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=max.LazyTensor]{max.LazyTensor()}}, else see
\code{\link[=max.default]{max.default()}}.
}
\examples{
\dontrun{
# R base operation
max(1:10)
max(c(NA, 1, 2), na.rm = TRUE)
# LazyTensor operation
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

max_x <- max(x_i)                   # LazyTensor object
max_red_x <- max(x_i, "i")          # max reduction indexed by 'i'
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
