% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{max_argmax}
\alias{max_argmax}
\alias{max_argmax_reduction}
\title{Max-ArgMax reduction}
\usage{
max_argmax(x, index)

max_argmax_reduction(x, index)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{a \code{character} corresponding to the wanted reduction dimension,
either \code{"i"} or \code{"j"}, to specify whether if the Max-ArgMax reduction is done
along the index \code{i} or \code{j}.}
}
\value{
an array storing the result of the specified Max-ArgMax
reduction otherwise.
}
\description{
Max-ArgMax reduction for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{max_argmax(x, index)} will :
\itemize{
\item if \code{index = "i"}, return the Max-ArgMax reduction of \code{x} over the \code{i}
indexes.
\item if \code{index = "j"}, return the Max-ArgMax reduction of \code{x} over the \code{j}
indexes.
}

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Reductions"}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

max_argmax_x <- max_argmax(x_i, "i")  # max argmax reduction indexed by 'i'
}
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

max_argmax_red <- max_argmax_reduction(x_i, "i")  # max argmax reduction 
                                                  # indexed by 'i'
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
