% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{min.LazyTensor}
\alias{min.LazyTensor}
\title{Minimum.}
\usage{
\method{min}{LazyTensor}(x, index = NA)
}
\description{
Minimum.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
