% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{min_argmin}
\alias{min_argmin}
\alias{min_argmin_reduction}
\title{Min-ArgMin reduction}
\usage{
min_argmin(x, index)

min_argmin_reduction(x, index)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{a \code{character} corresponding to the wanted reduction dimension,
either \code{"i"} or \code{"j"}, to specify whether if the Min-ArgMin reduction is done
along the index \code{i} or \code{j}.}
}
\value{
an array storing the result of the specified Min-ArgMin
reduction otherwise.
}
\description{
Min-ArgMin reduction for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{min_argmin(x, index)} will :
\itemize{
\item if \code{index = "i"}, return the Min-ArgMin reduction of \code{x}, i.e. the minimal
values and its position, over the \code{i} indexes.
\item if \code{index = "j"}, return the Min-ArgMin reduction of \code{x}, i.e. the minimal
values and its position, over the \code{j} indexes.
}

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Reductions"}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

min_argmin_xi <- min_argmin(x_i, "i")  # min argmin reduction indexed by 'i'
}
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

min_argmin_red <- min_argmin_reduction(x_i, "i") # min reduction indexed by 'i'
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
