% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Mod.LazyTensor}
\alias{Mod.LazyTensor}
\alias{Mod.ComplexLazyTensor}
\title{Element-wise complex modulus (absolute value)}
\usage{
\method{Mod}{LazyTensor}(z)

\method{Mod}{ComplexLazyTensor}(z)
}
\arguments{
\item{z}{A \code{ComplexLazyTensor} or any type of values accepted by R default
\code{\link[base:complex]{base::Mod()}} function.}
}
\value{
An object of class \code{ComplexLazyTensor}.
}
\description{
Symbolic element-wise complex modulus (absolute value) operation for
\code{ComplexLazyTensor} objects.
}
\details{
If \code{z} is a \code{ComplexLazyTensor}, \code{Mod(z)} returns a
\code{ComplexLazyTensor} that encodes, symbolically, the element-wise conjugate of
complex \code{z}.

\strong{Note}: \code{Mod(z)} will exit with an error if \code{z} is a \code{LazyTensor} but not
a \code{ComplexLazyTensor}.

\strong{Warning}: Do not confuse with \code{\link[=mod]{mod()}}.
}
\examples{
\dontrun{
z <- matrix(1i^ (-6:5), nrow = 4)                     # complex 4x3 matrix
z_i <- LazyTensor(z, index = 'i', is_complex = TRUE)  # ComplexLazyTensor
Mod_z_i <- Mod(z_i)                                   # symbolic matrix
}
}
\seealso{
\code{\link[=Mod]{Mod()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
