% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{mod}
\alias{mod}
\title{Element-wise modulo with offset function.}
\usage{
mod(x, ...)
}
\arguments{
\item{x}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values, or a scalar value.}

\item{a}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}

\item{b}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolic ternary operation for element-wise modulo with offset function.
}
\details{
\code{mod(x, a, b)} returns a \code{LazyTensor} that encodes, symbolically,
the element-wise modulo of \code{x} with modulus \code{a} and offset \code{b}. That is,
\code{mod(x, a, b)} encodes symbolically \code{x - a * floor((x - b)/a)}.
By default \code{b = 0}, so that \code{mod(x, a)} becomes equivalent to the R function \code{\%\%}.
\code{a} and \code{b} may be fixed integers or floats, or other \code{LazyTensor}.
Broadcasting rules apply.

\strong{Note}

If \code{a} and \code{b} are not scalar values, these should have the same inner dimension as \code{x}.

\strong{Warning}

Do not confuse with \code{Mod()}.
}
\examples{
\dontrun{
TODO add example
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
