% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Mod}
\alias{Mod}
\title{Element-wise complex modulus (absolute value) operation}
\usage{
Mod(z)
}
\arguments{
\item{z}{input for \code{\link[=Mod.default]{Mod.default()}} or \code{\link[=Mod.LazyTensor]{Mod.LazyTensor()}}.}
}
\value{
See value of \code{\link[=Mod.default]{Mod.default()}} or \code{\link[=Mod.LazyTensor]{Mod.LazyTensor()}}.
}
\description{
Symbolic element-wise complex modulus (absolute value) operation for
\code{ComplexLazyTensor} objects or default element-wise complex modulus
(absolute value) operation otherwise.
}
\details{
If \code{z} is a \code{LazyTensor}, see \code{\link[=Mod.LazyTensor]{Mod.LazyTensor()}}, else
see \code{\link[=Mod.default]{Mod.default()}}.

\strong{Warning}: Do not confuse with \code{\link[=mod]{mod()}}.
}
\examples{
\dontrun{
# R base operation
Mod(1)
Mod(1+1i)
# LazyTensor symbolic element-wise complex modulus (absolute value)
z <- matrix(1i^ (-6:5), nrow = 4)                     # complex 4x3 matrix
z_i <- LazyTensor(z, index = 'i', is_complex = TRUE)  # ComplexLazyTensor
Mod_z_i <- Mod(z_i)                                   # symbolic matrix
}
}
\seealso{
\code{\link[=Mod.default]{Mod.default()}}, \code{\link[=Mod.LazyTensor]{Mod.LazyTensor()}},
\code{\link[=Re]{Re()}}, \code{\link[=Im]{Im()}}, \code{\link[=Arg]{Arg()}}, \code{\link[=Conj]{Conj()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
