% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{mod.LazyTensor}
\alias{mod.LazyTensor}
\title{Element-wise modulo with offset function.}
\usage{
\method{mod}{LazyTensor}(x, a, b = 0, ...)
}
\description{
Element-wise modulo with offset function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
