% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-format_var_aliase.R
\name{parse_extra_args}
\alias{parse_extra_args}
\title{Parse formula for extra arguments not encoded in argument aliases}
\usage{
parse_extra_args(formula, args)
}
\arguments{
\item{formula}{text string, an operator formula (see Details).}

\item{args}{vector of text string, formula input arguments (see Details).}
}
\value{
a list with different information about formula input arguments:
a list with different information about formula input arguments:
\item{var_type}{vector of text string, corresponding type of formula
arguments (among \code{Vi}, \code{Vj}, \code{Pm}).}
\item{var_pos}{vector of integer, corresponding arguments positions.}
\item{var_dim}{vector of integer, corresponding arguments dimensions.}
}
\description{
FIXME
}
\details{
Parse the formula for string such as
\itemize{
\item \code{"YY(<pos>,<dim>)"} where \code{YY} can be a formula input argument type
(\code{Vi}, \code{Vj} or \code{Pm}), \verb{<pos>} is the position of the corresponding input
argument, and \verb{<dim>} its inner dimension.
\item \code{"Var(<pos>,<dim>,<type>)"} where \verb{<pos>} and \verb{<dim>} are the position and
inner dimension (c.f. previous point) and \verb{<type>} is an integer encoding
the formula input argument type with the following relation:
}

|---------|-------------------------|-----------|
| keyword | meaning                 | type      |
|---------|-------------------------|-----------|
| \code{Vi}    | variable indexed by \code{i} | \code{0}       |
| \code{Vj}    | variable indexed by \code{j} | \code{1}       |
| \code{Pm}    | parameter               | \code{2}       |
|---------|-------------------------|-----------|
}
\author{
Ghislain Durif
}
\keyword{internal}
