% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{Pm}
\alias{Pm}
\title{Wrapper LazyTensor parameter}
\usage{
Pm(x)
}
\arguments{
\item{x}{A vector or a scalar value.}
}
\value{
An object of class "LazyTensor" in parameter category.
}
\description{
Simple wrapper that return an instantiation of \code{LazyTensor} indexed by "i".
Equivalent to \code{LazyTensor(x)}.
}
\examples{
\dontrun{
x <- 4
Pm_x <- Pm(x)
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
