% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{Pm}
\alias{Pm}
\title{Wrapper LazyTensor parameter.}
\usage{
Pm(x, is_complex = FALSE)
}
\arguments{
\item{x}{A vector or a scalar value.}

\item{is_complex}{A boolean (default is FALSE). Whether if we want to create a
\code{ComplexLazyTensor} (is_complex = TRUE) or a \code{LazyTensor} (is_complex = FALSE).}
}
\value{
An object of class "LazyTensor" in parameter category. See \code{?LazyTensor} for more details.
}
\description{
Simple wrapper that return an instantiation of \code{LazyTensor} indexed by "i".
Equivalent to \code{LazyTensor(x)}.
}
\examples{
\dontrun{
x <- 4
Pm_x <- Pm(x)
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
