% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_preprocess.R
\name{Pm}
\alias{Pm}
\title{Wrapper LazyTensor parameter.}
\usage{
Pm(x, is_complex = FALSE)
}
\arguments{
\item{x}{A vector or a scalar value.}

\item{is_complex}{A boolean (default is FALSE). Whether if we want to create
a \code{ComplexLazyTensor} (\code{is_complex = TRUE}) or a \code{LazyTensor}
(\code{is_complex = FALSE}).}
}
\value{
An object of class "LazyTensor" in parameter category.
See \code{?LazyTensor} for more details.
}
\description{
Simple wrapper returning an instance of a fixed parameter \code{LazyTensor}.
Equivalent to \code{LazyTensor(x)}.
}
\details{
See \code{?LazyTensor} for more details.
}
\examples{
\dontrun{
x <- 4
Pm_x <- Pm(x)
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
