% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Re.LazyTensor}
\alias{Re.LazyTensor}
\title{Element-wise real part of complex.}
\usage{
\method{Re}{LazyTensor}(z)
}
\description{
Element-wise real part of complex.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
