% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{Re}
\alias{Re}
\title{Element-wise complex real part operation}
\usage{
Re(z)
}
\arguments{
\item{z}{input for \code{\link[=Re.default]{Re.default()}} or \code{\link[=Re.LazyTensor]{Re.LazyTensor()}}.}
}
\value{
See value of \code{\link[=Re.default]{Re.default()}} or \code{\link[=Re.LazyTensor]{Re.LazyTensor()}}.
}
\description{
Symbolic element-wise complex real part operation for \code{ComplexLazyTensor}
objects or default R element-wise complex real part operation otherwise.
}
\details{
If \code{z} is a \code{ComplexLazyTensor}, see
\code{\link[=Re.LazyTensor]{Re.LazyTensor()}}, else see \code{\link[=Re.default]{Re.default()}}.
}
\examples{
\dontrun{
# R base operation
Re(1)
Re(1+1i)
# LazyTensor symbolic element-wise complex real part
z <- matrix(2 + 1i^ (-6:5), nrow = 4)        # complex matrix
z_i <- LazyTensor(z, "i", is_complex = TRUE) # creating ComplexLazyTensor
Re_z <- Re(z_i)                              # symbolic matrix
}
}
\seealso{
\code{\link[=Re.default]{Re.default()}}, \code{\link[=Re.LazyTensor]{Re.LazyTensor()}},
\code{\link[=Im]{Im()}}, \code{\link[=Arg]{Arg()}}, \code{\link[=Mod]{Mod()}}, \code{\link[=Conj]{Conj()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
