% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{reduction.LazyTensor}
\alias{reduction.LazyTensor}
\title{Reduction operation.}
\usage{
reduction.LazyTensor(x, opstr, index)
}
\arguments{
\item{x}{A \code{LazyTensor}, a vector or a matrix of numeric values, or a scalar value.}

\item{opstr}{A \code{string} formula (like "Sum" or "Max").}

\item{index}{A \code{character} that should be either \strong{i} or \strong{j} to specify whether if
the reduction is indexed by \strong{i} (rows), by \strong{j} (columns).}
}
\value{

}
\description{
Applies a reduction to a \code{LazyTensor}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'

red_x <- reduction.LazyTensor(x_i, "Sum", "i")
}
}
\author{
Ghislain Durif
}
