\name{rkeops-package}
\alias{rkeops-package}
\alias{rkeops}
\docType{package}
\title{
\packageTitle{rkeops}
}
\description{
\packageDescription{rkeops}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rkeops}
\packageIndices{rkeops}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rkeops}

Maintainer: \packageMaintainer{rkeops}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
