% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops-package.R
\docType{package}
\name{rkeops-package}
\alias{rkeops-package}
\alias{rkeops}
\title{rkeops

RKeOps: Kernel Operations on the GPU, with autodiff, without memory overflows

RKeOps contains the R bindings for the cpp/cuda library
\href{https://www.kernel-operations.io/}{KeOps}. It provides
standard R functions that can be used in any R (>=3) codes.

For a full documentation you may read:
\itemize{
\item \href{https://www.kernel-operations.io/keops/introduction/installation.html}{Installation}
\item \href{https://www.kernel-operations.io/}{Documentation}
\item \href{https://www.kernel-operations.io/keops/_auto_examples/index.html}{Learning KeOps syntax with examples}
\item \href{https://www.kernel-operations.io/keops/_auto_tutorials/index.html}{Tutorials gallery}
}}
\description{
The KeOps library lets you compute generic reductions of very large arrays whose
entries are given by a mathematical formula. It combines a tiled reduction scheme
with an automatic differentiation engine, and can be used through Matlab, NumPy or
PyTorch backends. It is perfectly suited to the computation of Kernel dot products
and the associated gradients, even when the full kernel matrix does not fit into
the GPU memory.
}
\details{
See \url{https://www.kernel-operations.io/}
}
\author{
\itemize{
\item \href{http://imag.umontpellier.fr/~charlier/}{Benjamin Charlier}
\item \href{https://www.math.ens.fr/~feydy/}{Jean Feydy}
\item \href{https://www.mi.parisdescartes.fr/~glaunes/}{Joan Alexis Glaunès}
\item \href{https://gdurif.perso.math.cnrs.fr/}{Ghislain Durif}
}
}
