% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_options.R
\name{rkeops_disable_verbosity}
\alias{rkeops_disable_verbosity}
\title{Disable additional verbosity in \code{rkeops}}
\usage{
rkeops_disable_verbosity()
}
\value{
None
}
\description{
Disable verbosity during operator compilation process.
}
\examples{
\dontrun{
rkeops_disable_verbosity()
}
}
\seealso{
\code{\link[=rkeops_enable_verbosity]{rkeops_enable_verbosity()}}, \code{\link[=set_rkeops_options]{set_rkeops_options()}}
}
\author{
Ghislain Durif
}
