% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_options.R
\name{rkeops_enable_verbosity}
\alias{rkeops_enable_verbosity}
\title{Enable additional verbosity in \code{rkeops}}
\usage{
rkeops_enable_verbosity()
}
\value{
None
}
\description{
Enable verbosity during operator compilation process.
}
\examples{
\dontrun{
rkeops_enable_verbosity()
}
}
\seealso{
\code{\link[=rkeops_disable_verbosity]{rkeops_disable_verbosity()}}, \code{\link[=set_rkeops_options]{set_rkeops_options()}}
}
\author{
Ghislain Durif
}
