% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{round.default}
\alias{round.default}
\title{Rounding function.}
\usage{
\method{round}{default}()
}
\description{
Rounding function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
