% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{round.LazyTensor}
\alias{round.LazyTensor}
\title{Element-wise rounding function.}
\usage{
\method{round}{LazyTensor}(x, d)
}
\description{
Element-wise rounding function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
