% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{round}
\alias{round}
\title{Element-wise rounding function}
\usage{
round(x, digits = 0)
}
\arguments{
\item{x}{input for \code{\link[=round.default]{round.default()}} or \code{\link[=round.LazyTensor]{round.LazyTensor()}}.}

\item{digits}{integer indicating the number of decimal places to be used
when rounding.}
}
\value{
See value of \code{\link[=round.default]{round.default()}} or \code{\link[=round.LazyTensor]{round.LazyTensor()}}.
}
\description{
Symbolic element-wise rounding function for \code{LazyTensor} objects or
default element-wise rounding function otherwise.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=round.LazyTensor]{round.LazyTensor()}}, else
see \code{\link[=round.default]{round.default()}}.
}
\examples{
\dontrun{
# R base operation
round(5)
round(runif(10), 4)
# LazyTensor symbolic element-wise rounding
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Round_x <- round(x_i, 2)            # symbolic matrix
}
}
\seealso{
\code{\link[=round.default]{round.default()}}, \code{\link[=round.LazyTensor]{round.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
