% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{rsqrt.LazyTensor}
\alias{rsqrt.LazyTensor}
\title{Element-wise inverse square root.}
\usage{
\method{rsqrt}{LazyTensor}(x)
}
\description{
Element-wise inverse square root.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
