% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{rsqrt}
\alias{rsqrt}
\title{Element-wise inverse square root operation}
\usage{
rsqrt(x)
}
\arguments{
\item{x}{input for \code{\link[=rsqrt.default]{rsqrt.default()}} or
\code{\link[=rsqrt.LazyTensor]{rsqrt.LazyTensor()}}.}
}
\value{
See value of \code{\link[=rsqrt.default]{rsqrt.default()}} or
\code{\link[=rsqrt.LazyTensor]{rsqrt.LazyTensor()}}.
}
\description{
Symbolic element-wise inverse square root operation for \code{LazyTensor}
objects or standard element-wise inverse square root operation otherwise,
i.e. \eqn{1/\sqrt{x}}.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=rsqrt.LazyTensor]{rsqrt.LazyTensor()}}, else
see \code{\link[=rsqrt.default]{rsqrt.default()}}.
}
\examples{
\dontrun{
# Numerical input
rsqrt(4)
rsqrt(c(1,4,9,16))
# LazyTensor symbolic element-wise inverse square root
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Rsqrt_x <- rsqrt(x_i)               # symbolic matrix, 150 rows and 3 columns
}
}
\seealso{
\code{\link[=rsqrt.default]{rsqrt.default()}}, \code{\link[=rsqrt.LazyTensor]{rsqrt.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
