% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{scalar.product.LazyTensor}
\alias{scalar.product.LazyTensor}
\alias{|.LazyTensor}
\title{Euclidean scalar product for LazyTensors.}
\usage{
\method{|}{LazyTensor}(x, y)
}
\description{
Euclidean scalar product for LazyTensors.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
