% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-set_options.R
\name{set_rkeops_options}
\alias{set_rkeops_options}
\title{Initialize or update \code{rkeops} options in \code{R} global options scope}
\usage{
set_rkeops_options(custom_compile_options = NULL,
  custom_runtime_options = NULL)
}
\arguments{
\item{custom_compile_options}{a list (of class \code{rkeops_compile_options})
with the following elements \code{build_dir}, \code{src_dir}, \code{precision}, \code{verbosity},
\code{use_cuda_if_possible}. Default value is \code{NULL} and default compile options
are set up (see \code{\link[rkeops:default_compile_options]{rkeops::default_compile_options()}}).}

\item{custom_runtime_options}{a list (of class \code{rkeops_runtime_options})
with the following elements \code{tagCpuGpu}, \code{tag1D2D}, \code{device_id}. Default
value is \code{NULL} and default runtime options are set up
(see \code{\link[rkeops:default_runtime_options]{rkeops::default_runtime_options()}}).}
}
\description{
\code{rkeops} uses two sets of options: compile options (see
\code{\link[rkeops:compile_options]{rkeops::compile_options()}}) and runtime options (see
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}). These options define the behavior of \code{rkeops}
when compiling or when running new user-defined operators.

If no input is provided, the functions \code{set_rkeops_options} initializes the
\code{rkeops} options in the \code{R} global options scope (i.e. options available
by calling \code{options()} or \code{getOptions(<option_name>)}).

If some input is provided, i.e. objects defining compile options or run
options (see Details), \code{rkeops} global options are updated accordingly.
}
\details{
\code{rkeops} global options includes two lists defining options used at
compilation of new user-defined operators or at runtime. These two list
contains specific informations (see \code{\link[rkeops:compile_options]{rkeops::compile_options()}} and
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}} respectively).

In order to update, the corresponding options, user should provide objects
returned by the functions \code{\link[rkeops:compile_options]{rkeops::compile_options()}} and
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}} respectively, being lists of class
\code{rkeops_compile_options} and \code{rkeops_runtime_options} respectively, with
specific attributes.
}
\seealso{
\code{\link[rkeops:set_rkeops_option]{rkeops::set_rkeops_option()}}, \code{\link[rkeops:compile_options]{rkeops::compile_options()}},
\code{\link[rkeops:runtime_options]{rkeops::runtime_options()}}, \code{\link[rkeops:default_compile_options]{rkeops::default_compile_options()}},
\code{\link[rkeops:default_runtime_options]{rkeops::default_runtime_options()}}.
}
\author{
Ghislain Durif
}
