% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sign.LazyTensor}
\alias{sign.LazyTensor}
\title{Element-wise sign.}
\usage{
\method{sign}{LazyTensor}(x)
}
\description{
Element-wise sign.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
