% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sin.LazyTensor}
\alias{sin.LazyTensor}
\title{Element-wise sine.}
\usage{
\method{sin}{LazyTensor}(x)
}
\description{
Element-wise sine.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
