% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sinxdivx.LazyTensor}
\alias{sinxdivx.LazyTensor}
\title{Element-wise sin(x)/x function.}
\usage{
\method{sinxdivx}{LazyTensor}(x)
}
\description{
Element-wise sin(x)/x function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
