% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sqrt.LazyTensor}
\alias{sqrt.LazyTensor}
\title{Element-wise square root.}
\usage{
\method{sqrt}{LazyTensor}(x)
}
\description{
Element-wise square root.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
