% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sqrt}
\alias{sqrt}
\title{Element-wise square root operation}
\usage{
sqrt(x)
}
\arguments{
\item{x}{input for \code{\link[=sqrt.default]{sqrt.default()}} or \code{\link[=sqrt.LazyTensor]{sqrt.LazyTensor()}}.}
}
\value{
See value of \code{\link[=sqrt.default]{sqrt.default()}} or \code{\link[=sqrt.LazyTensor]{sqrt.LazyTensor()}}.
}
\description{
Symbolic element-wise square root operation for \code{LazyTensor} objects or
standard element-wise square root operation otherwise, i.e.
\eqn{\sqrt{x}}.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=sqrt.LazyTensor]{sqrt.LazyTensor()}}, else
see \code{\link[=sqrt.default]{sqrt.default()}}.

\strong{Note}: for complex input argument \code{z}, then the absolute value
corresponds to the complex modulus \code{sqrt(z) = z^0.5}.
}
\examples{
\dontrun{
# R base operation
sqrt(4)
sqrt(c(1,4,9,16))
# LazyTensor symbolic element-wise square root
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Sqrt_x <- sqrt(x_i)                 # symbolic matrix, 150 rows and 3 columns
}
}
\seealso{
\code{\link[=sqrt.default]{sqrt.default()}}, \code{\link[=sqrt.LazyTensor]{sqrt.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
