% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{square.LazyTensor}
\alias{square.LazyTensor}
\title{Element-wise square.}
\usage{
\method{square}{LazyTensor}(x)
}
\description{
Element-wise square.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
