% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_utils.R
\name{stat_rkeops_cache_dir}
\alias{stat_rkeops_cache_dir}
\title{Disk usage of RKeOps cache directory content}
\usage{
stat_rkeops_cache_dir(verbose = TRUE, startup = FALSE)
}
\arguments{
\item{verbose}{logical, indicates whether a message about rkeops cache
directory disk usage should be printed or not.}

\item{startup}{logical, internal use for verbosity at package loading.}
}
\value{
string character, disk usage in bytes of the rkeops cache
directory.
}
\description{
Get the disk usage in bytes for all files in the folder where rkeops operator
compilation byproducts are stored to be re-used for further use
(and avoid unnecessary recompilation).
}
\details{
When compiling a user-defined operators or running a LazyTensor-based
operations, a shared object (\code{.so}) library (or dynamic link library,
\code{.dll}) file is created in RKeOps cache directory.

For every new operators or new LazyTensor operations, such a file is
created, to be reused next time the same computations is done (then saving
compilation time).

Calling \code{stat_rkeops_cache_dir()} gives RKeOps disk usage in bytes.
cache directory.

You can use \code{\link[=get_rkeops_cache_dir]{get_rkeops_cache_dir()}} to get the path to RKeOps cache
directory, and you can use \code{\link[=clean_rkeops]{clean_rkeops()}} to (fully or partially)
delete its content.
}
\examples{
\dontrun{
stat_rkeops_cache_dir()
}
}
\seealso{
\code{\link[=get_rkeops_cache_dir]{get_rkeops_cache_dir()}}, \code{\link[=clean_rkeops]{clean_rkeops()}}
}
\author{
Ghislain Durif
}
