% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{step.LazyTensor}
\alias{step.LazyTensor}
\title{Element-wise step function.}
\usage{
\method{step}{LazyTensor}(x)
}
\description{
Element-wise step function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
