% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{step}
\alias{step}
\title{Element-wise 0-1 step function (for LazyTensors) or default stepwise model
selection}
\usage{
step(object, ...)
}
\arguments{
\item{object}{input for \code{\link[=step.default]{step.default()}} or
\code{\link[=step.LazyTensor]{step.LazyTensor()}}.}

\item{...}{optional additional input arguments.}
}
\value{
See value of \code{\link[=step.default]{step.default()}} or \code{\link[=step.LazyTensor]{step.LazyTensor()}}.
}
\description{
Symbolic element-wise 0-1 step function for \code{LazyTensor} objects
(i.e. \code{0} if \code{x < 0}, \code{1} if \code{x >= 0} for an input \code{x}) or stepwise
model selection otherwise.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=step.LazyTensor]{step.LazyTensor()}}, else
see \code{\link[=step.default]{step.default()}}.
}
\examples{
\dontrun{
# R base operation: see `?step.default`
# LazyTensor symbolic element-wise sign
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
Step_x <- step.LazyTensor(x_i)      # symbolic matrix, 150 rows and 3 columns
}
}
\seealso{
\code{\link[=step.default]{step.default()}}, \code{\link[=step.LazyTensor]{step.LazyTensor()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
