% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sum.LazyTensor}
\alias{sum.LazyTensor}
\title{Summation operation or Sum reduction.}
\usage{
\method{sum}{LazyTensor}(x, index = NA)
}
\description{
Summation operation or Sum reduction.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
