% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sum.LazyTensor}
\alias{sum.LazyTensor}
\title{Sum operation or sum reduction}
\usage{
\method{sum}{LazyTensor}(x, index = NA, ...)
}
\arguments{
\item{x}{a \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector or a matrix of
numeric values, or a scalar value.}

\item{index}{a \code{character} corresponding to the wanted reduction dimension,
either \code{i} or \code{j}, to specify whether if the sum reduction is done along the
index \code{i}, or \code{j}.
It can be \code{NA} (default) when no reduction is desired.}
}
\value{
a \code{LazyTensor} if \code{index = NA} or an array storing the result of
the specified sum reduction otherwise.
}
\description{
Sum operation or Sum reduction for \code{LazyTensor} objects.
}
\details{
If \code{x} is a \code{LazyTensor}, \code{sum(x, index)} will :
\itemize{
\item if \code{index = "i"}, return the sum reduction of \code{x} over the \code{"i"}
indexes.
\item if \code{index = "j"}, return the sum reduction of \code{x} over the \code{"j"}
indexes.
\item if \code{index = NA} (default), return a new \code{LazyTensor} object
representing the symbolic sum of the values along the inner dimension of
\code{x}.
}

\code{Notes}: If \code{index = NA}, \code{x} input argument should be a \code{LazyTensor}
encoding a parameter vector.

Run \code{browseVignettes("rkeops")} to access the vignettes and find details
about the function in "RKeOps LazyTensor" vignette, at section "Reductions".
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

sum_x <- sum(x_i)                   # LazyTensor object
sum_red_x <- sum(x_i, "i")          # sum reduction indexed by 'i'
}

}
\author{
Chloe Serre-Combe, Amelie Vernay
}
