% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{sum}
\alias{sum}
\title{Summation operation or Sum reduction.}
\usage{
sum(x, index)
}
\arguments{
\item{x}{A \code{LazyTensor}, a vector or a matrix of numeric values, or a scalar value.}

\item{index}{A \code{character} that should be either \strong{i} or \strong{j} or \strong{NA} to specify whether if
the summation is indexed by \strong{i} (rows), by \strong{j} (columns).}
}
\value{

}
\description{
Summation unary operation, or Sum reduction.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'

sum_x <- sum(x_i) # LazyTensor object
sum_red_x <- sum(x_i, "i")  # sum reduction indexed by 'i'
}
}
\author{
Ghislain Durif
}
