% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sum}
\alias{sum}
\title{Sum operation or Sum reduction.}
\usage{
sum(x, ...)
}
\arguments{
\item{x}{input for \code{\link[=sum.default]{sum.default()}} or \code{\link[=sum.LazyTensor]{sum.LazyTensor()}}.}

\item{...}{optional arguments for \code{\link[=sum.default]{sum.default()}} or
\code{\link[=sum.LazyTensor]{sum.LazyTensor()}}.}
}
\value{
Sum of input.
}
\description{
Sum operation for vector or LazyTensor, or Sum reduction for
LazyTensor.
}
\details{
If \code{x} is a \code{LazyTensor}, see \code{\link[=sum.LazyTensor]{sum.LazyTensor()}}, else see
\code{\link[=sum.default]{sum.default()}}.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
