% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{sum_reduction}
\alias{sum_reduction}
\title{Sum reduction.}
\usage{
sum_reduction(x, index)
}
\arguments{
\item{x}{a \code{LazyTensor} or a \code{ComplexLazyTensor}.}

\item{index}{a \code{character} corresponding to the wanted reduction dimension,
either \code{i} or \code{j}, to specify whether if the sum reduction is done along the
index \code{i}, or \code{j}.}
}
\value{
an array storing the result of the specified sum reduction.
}
\description{
Sum reduction for \code{LazyTensor} objects.
}
\details{
\code{sum_reduction(x, index)} will return the sum reduction of \code{x}
over the given \code{index}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'

sum_reduction(x_i, "i")
}
}
\seealso{
\code{\link[=sum]{sum()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
