% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{tensorprod}
\alias{tensorprod}
\title{Tensor product.}
\usage{
tensorprod(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{

}
\description{
Tensor product of vectors - a binary operation.
}
\examples{
\dontrun{
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
