% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{ternaryop.LazyTensor}
\alias{ternaryop.LazyTensor}
\title{Build a ternary operation}
\usage{
ternaryop.LazyTensor(x, y, z, opstr)
}
\arguments{
\item{x}{A LazyTensor, a vector of numeric values, or a scalar value.}

\item{y}{A LazyTensor, a vector of numeric values, or a scalar value.}

\item{opstr}{A text string corresponding to an operation.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolically applies \strong{opstr} operation to \strong{x}, \strong{y} and \strong{z}.
}
\examples{
\dontrun{
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
