% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{ternaryop.LazyTensor}
\alias{ternaryop.LazyTensor}
\title{Build a ternary operation}
\usage{
ternaryop.LazyTensor(x, y, z, opstr, dim_check_type = "sameor1")
}
\arguments{
\item{x}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values, or a scalar value.}

\item{y}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values, or a scalar value.}

\item{z}{A \code{LazyTensor}, a \code{ComplexLazyTensor}r, a vector of numeric values, or a scalar value.}

\item{opstr}{A text string corresponding to an operation.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolically applies \strong{opstr} operation to \strong{x}, \strong{y} and \strong{z}.
}
\examples{
\dontrun{
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
