% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{unaryop.LazyTensor}
\alias{unaryop.LazyTensor}
\title{Build a unary operation}
\usage{
unaryop.LazyTensor(x, opstr, opt_arg = NA, opt_arg2 = NA)
}
\arguments{
\item{x}{A LazyTensor, a vector of numeric values, or a scalar value.}

\item{opstr}{A text string corresponding to an operation.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolically applies \strong{opstr} operation to \strong{x}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'
una_xy <- unaryop.LazyTensor(x_i, "Minus")   # symbolic matrix

una2_xy <- unaryop.LazyTensor(x_i, "Pow", opt_arg = 3)  # symbolic matrix
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
