% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{unaryop.LazyTensor}
\alias{unaryop.LazyTensor}
\title{Build a unary operation}
\usage{
unaryop.LazyTensor(x, opstr)
}
\arguments{
\item{x}{A LazyTensor, a vector of numeric values, or a scalar value.}

\item{opstr}{A text string corresponding to an operation.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolically applies \strong{opstr} operation to \strong{x}.
}
\examples{
\dontrun{
}
}
\author{
Ghislain Durif
}
