% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{vecmatmult}
\alias{vecmatmult}
\title{Vector-matrix product.}
\usage{
vecmatmult(x, y)
}
\arguments{
\item{x}{D An \code{integer} corresponding to the output dimension.}
}
\value{

}
\description{
Vector-matrix product - a binary operation.
}
\details{
vector-matrix product \code{x} x \code{y}: \code{x} is vector, \code{y} is vector
interpreted as matrix (column-major)
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # LazyTensor from matrix x, indexed by 'i'


mv_mult_xy <- matvecmultt(x_i, y_j) # symbolic matrix
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
