% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{Vi}
\alias{Vi}
\title{Wrapper LazyTensor indexed by "i"}
\usage{
Vi(x)
}
\arguments{
\item{x}{A matrix of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor" indexed by "i".
}
\description{
Simple wrapper that return an instantiation of \code{LazyTensor} indexed by "i".
Equivalent to \code{LazyTensor(x, index="i")}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3)
Vi_x <- Vi(x)
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
