% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{Vj}
\alias{Vj}
\title{Wrapper LazyTensor indexed by "j"}
\usage{
Vj(x)
}
\arguments{
\item{x}{A matrix of numeric values.}
}
\value{
An object of class "LazyTensor" indexed by "j".
}
\description{
Simple wrapper that return an instantiation of \code{LazyTensor} indexed by "j".
Equivalent to \code{LazyTensor(x, index="j")}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3)
Vj_x <- Vj(x)
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
