% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{weightedsqdist}
\alias{weightedsqdist}
\title{Generic squared distance.}
\usage{
weightedsqdist(x, y, z)
}
\arguments{
\item{x}{A \code{vector} of numeric values or a scalar value.}

\item{y}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}

\item{z}{A \code{LazyTensor}, a vector of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolic binary operation for weighted squared distance of a LazyTensor.
}
\details{
\code{weightedsqdist(x)} returns a \code{LazyTensor} that encodes, symbolically,
the weighted squared distance of a vector \code{x} with weights stored in the LazyTensor \code{s}.
}
\examples{
\dontrun{
}
}
\author{
Chloé Serre-Combe, Amélie Vernay
}
