% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkeops_lazytensor_operations.R
\name{weightedsqnorm}
\alias{weightedsqnorm}
\title{Generic squared eucidian norm.}
\usage{
weightedsqnorm(x, s)
}
\arguments{
\item{x}{A \code{vector} of numeric values or a scalar value.}

\item{s}{A \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values, or a scalar value.}
}
\value{
An object of class "LazyTensor".
}
\description{
Symbolic binary operation for weighted squared norm of a LazyTensor.
}
\details{
\code{weightedsqnorm(x)} returns a \code{LazyTensor} that encodes, symbolically,
the weighted squared norm of a vector \code{x} with weights stored in the LazyTensor \code{s}.
}
\examples{
\dontrun{
x <- matrix(runif(150 * 3), 150, 3) # arbitrary R matrix, 150 rows and 3 columns
y <- matrix(runif(100 * 4), 100, 4) # arbitrary R matrix, 100 rows and 4 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, indexed by 'i'
y_j <- LazyTensor(y, index = 'j')   # creating LazyTensor from matrix y, indexed by 'j'

wsqn_xy <- weightedsqnorm(x_i, y_j)
}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
