% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{weightedsqnorm}
\alias{weightedsqnorm}
\title{Generic weighted squared Euclidean norm}
\usage{
weightedsqnorm(x, s)
}
\arguments{
\item{x}{a \code{vector} of numeric values or a scalar value.}

\item{s}{a \code{LazyTensor}, a \code{ComplexLazyTensor}, a vector of numeric values,
or a scalar value.}
}
\value{
An object of class \code{LazyTensor}.
}
\description{
Symbolic weighted squared norm of a LazyTensor binary operation for
\code{LazyTensor} objects.
}
\details{
\code{weightedsqnorm(x, s)} returns a \code{LazyTensor} that encodes,
symbolically, the weighted squared norm of a vector \code{x} with weights stored
in the LazyTensor \code{s}.

\strong{Note}: \code{x}, and \code{s} should have the same inner dimension.

\strong{Note}: Run \code{browseVignettes("rkeops")} to access the vignettes and find
details about this function in the \emph{"RKeOps LazyTensor"} vignette, at
section \emph{"Simple vector operations"}.
}
\examples{
\dontrun{
x <- matrix(runif(100 * 3), 100, 3) # arbitrary R matrix, 100 rows, 3 columns
s <- matrix(runif(100 * 3), 100, 3) # arbitrary R matrix, 100 rows, 3 columns
x_i <- LazyTensor(x, index = 'i')   # creating LazyTensor from matrix x, 
                                    # indexed by 'i'
s_j <- LazyTensor(s, index = 'j')   # creating LazyTensor from matrix s, 
                                    # indexed by 'j'
wsqn_xy <- weightedsqnorm(x_i, s_j) # symbolic matrix, 100 rows,1 columns
}
}
\seealso{
\code{\link[=sqnorm2]{sqnorm2()}}, \code{\link[=weightedsqdist]{weightedsqdist()}}
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
