% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytensor_operations.R
\name{xlogx.LazyTensor}
\alias{xlogx.LazyTensor}
\title{Element-wise x*log(x) function.}
\usage{
\method{xlogx}{LazyTensor}(x)
}
\description{
Element-wise x*log(x) function.
}
\author{
Chloe Serre-Combe, Amelie Vernay
}
\keyword{internal}
